% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHelpers.R
\name{add_dose}
\alias{add_dose}
\title{Add dose tally to table}
\usage{
add_dose(p, dose_level)
}
\arguments{
\item{dose_level}{"l" or "h" indicating whether low or high doses should be tallied in tables and plots}
}
\value{
table with dose level tally
}
\description{
Add dose tally to table
}
