% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBalStats.R
\name{calcBalStats}
\alias{calcBalStats}
\title{Code to calculate balance stats based on Jackson paper (either weighted or unweighted)}
\usage{
calcBalStats(
  data,
  forms,
  f_type,
  exposure,
  outcome,
  type,
  k = 1,
  bal_thresh = 0.1
)
}
\arguments{
\item{forms}{from createForms(), createShortForms, or updateForms}

\item{f_type}{form type label}

\item{exposure}{exposure}

\item{outcome}{outcome}

\item{k}{imputaiton number}

\item{object}{msm object that contains all relevant user inputs}

\item{imp_wide_data}{wide data}

\item{weighted}{binary indicator of whether to calculate weighted balance stats}

\item{histories}{binary indicator of whether to print history sample distributions}
}
\description{
Code to calculate balance stats based on Jackson paper (either weighted or unweighted)
}
\examples{
calcBalStats(object, imp_wide_data, forms, f_type, exposure, outcome, k=1, weighted=0, histories=1)
Add information to balance statistics (bal_stats)
}
