% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condenseWeights.R
\name{condenseWeights}
\alias{condenseWeights}
\title{condenseWeights}
\usage{
condenseWeights(ID, home_dir, m, weights_models, exposures, time_pts)
}
\arguments{
\item{ID}{person-level identifier in your dataset}

\item{home_dir}{path to home directory for the project}

\item{m}{number of imputed datasets from Amelia}

\item{weights_models}{output from createWeights}

\item{exposures}{list of variables that represent your exposures/treatments of interest}

\item{time_pts}{list of time points along your developmental path of interest for which you have at least one measurement}
}
\value{
data_for_model_with_weights
}
\description{
condenseWeights
}
