% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createForms.R
\name{createForms}
\alias{createForms}
\title{Creates formulas for generating weights}
\usage{
createForms(object, wide_long_datasets, all_potential_covariates)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{wide_long_datasets}{from formatForWeights}

\item{all_potential_covariates}{from selectCoariates}
}
\value{
forms
}
\description{
Creates full formulas for each exposure at each time point including all time-varying covariates at all prior lags
}
\examples{
createForms(object, wide_long_datasets, all_potential_covariates)

}
\seealso{
\code{\link[=formatForWeights]{formatForWeights()}}, \link{identifyPotentialConfounds} for more on inputs
}
