% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFormulas.R
\name{createFormulas}
\alias{createFormulas}
\title{Create balancing formulas}
\usage{
createFormulas(
  home_dir,
  exposure,
  exposure_time_pts,
  outcome,
  tv_confounders,
  ti_confounders,
  type,
  bal_stats = NULL,
  concur_conf = NULL,
  keep_conf = NULL,
  custom = NULL,
  verbose = TRUE,
  save.out = TRUE
)
}
\arguments{
\item{home_dir}{path to home directory}

\item{exposure}{name of exposure variable}

\item{exposure_time_pts}{list of integers at which weights will be
created/assessed that correspond to time points when exposure wass measured}

\item{outcome}{name of outcome variable with ".timepoint" suffix}

\item{tv_confounders}{list of time-varying confounders with ".timepoint"
suffix}

\item{ti_confounders}{list of time invariant confounders}

\item{type}{type of formula to create from 'full' (includes all lagged
time-varying confounders), 'short' (includes time-varying confounders at
t-1 lag only), or 'update' (adds to 'short' formulas any imbalanced
time-varying confounders at lags great than t-1)}

\item{bal_stats}{list of balance statistics from assessBalance(), required
for 'update' type}

\item{concur_conf}{(optional) list of variable names reflecting time-varying
confounders to retain in formulas contemporaneously (default is none)}

\item{keep_conf}{(optional) list of variable names reflecting confounders to
always retain in formulas (default depends on type)}

\item{custom}{(optional) custom list of formulas at each exposure time point
(default is to create automatically according to type)}

\item{verbose}{(optional) TRUE or FALSE indicator for user output (default is
TRUE)}

\item{save.out}{(optional) TRUE or FALSE indicator to save output and intermediary output locally (default is TRUE)}
}
\value{
list of balancing formulas at each exposure time point
}
\description{
Creates balancing formulas relating exposure to all relevant time-varying and
time invariant confounders at each exposure time point to be used to create
IPTW weights.
}
\examples{
test <- data.frame(ID = 1:10,
                   A.1 = 1:10,
                   A.2 = 21:30,
                   A.3 = 1:10,
                   B.1 = 2:11,
                   B.2 = 1:10,
                   B.3 = 4:13,
                   C = 3:12,
                   D.3 = 4:13)
createFormulas(getwd(), "A", c(1, 2, 3), "D.3", c("B.1", "B.2", "B.3"), "C", "full")
}
