% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWeights.R
\name{createWeights}
\alias{createWeights}
\alias{print.devMSM_weights}
\alias{plot.devMSM_weights}
\alias{summary.devMSM_weights}
\title{Creates IPTW balancing weights}
\usage{
createWeights(
  data,
  formulas,
  method = "glm",
  verbose = FALSE,
  save.out = FALSE,
  ...
)

\method{print}{devMSM_weights}(x, i = 1, ...)

\method{plot}{devMSM_weights}(x, i = 1, save.out = FALSE, ...)

\method{summary}{devMSM_weights}(object, i = 1, ...)
}
\arguments{
\item{data}{data in wide format as: a data frame, list of imputed data
frames, or \code{mids} object from the \code{mice} package}

\item{formulas}{list of balancing formulas at each time point output from \code{\link[=createFormulas]{createFormulas()}}}

\item{method}{character string of \code{weightitMSM()} balancing method
abbreviation (default is generalized linear models propensity score weighting \code{"glm"})}

\item{verbose}{(optional) TRUE or FALSE indicator for printing output to console.
default is FALSE.}

\item{save.out}{(optional) Either logical or a character string. If \code{TRUE},
it will output the result to a default file name within \code{home_dir} set in \code{initMSM()}. You can load the data with \code{x <- readRDS(file)}.
To use a non-default file name, specify a character string with the file name. It will save relative to \code{home_dir}.
There might be naming conflicts where two objects get saved to the same file. In these cases, users should specify a custom name.
default is FALSE.}

\item{\dots}{arguments passed to \code{\link[WeightIt:weightitMSM]{WeightIt::weightitMSM()}} or \code{\link[WeightIt:summary.weightit]{WeightIt::summary.weightitMSM()}}}

\item{x, object}{\code{devMSM_weights} object from \code{createWeights()}}

\item{i}{For multiply imputed datasets, \code{i} selects which imputation to print results for.
Default is \code{i = 1}. With \code{i = TRUE}, all imputed datasets will be looped over. With \code{i = NULL}, will average over all imputed datasets and summarize that. Ignored for non-imputed data.}
}
\value{
a list containing \code{\link[WeightIt:weightitMSM]{WeightIt::weightitMSM()}} output. It is the length
of the number of datasets (1 for a data.frame or the number of imputed datasets).
}
\description{
Creates IPTW balancing weights at each user-specified exposure time point
using balancing formulas that relate exposure at each time point to all
relevant confounders.
}
\examples{
library(devMSMs)
data <- data.frame(
  ID = 1:50,
  A.1 = rnorm(n = 50),
  A.2 = rnorm(n = 50),
  A.3 = rnorm(n = 50),
  B.1 = rnorm(n = 50),
  B.2 = rnorm(n = 50),
  B.3 = rnorm(n = 50),
  C = rnorm(n = 50),
  D.3 = rnorm(n = 50)
)
obj <- initMSM(
  data,
  exposure = c("A.1", "A.2", "A.3"),
  ti_conf = c("C"),
  tv_conf = c("B.1", "B.2", "B.3", "D.3")
)
f <- createFormulas(obj, type = "short")

w <- createWeights(data = data, formulas = f)
print(w)
plot(w)

# Methods from `WeightIt::weightitMSM`
w <- createWeights(data = data, formulas = f,
                   method = "glm")

w <- createWeights(data = data, formulas = f,
                   method = "cbps")
\dontshow{if (requireNamespace("gbm", quietly = TRUE)) withAutoprint(\{ # examplesIf}
w <- createWeights(data = data, formulas = f,
                   method = "gbm")
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("dbarts", quietly = TRUE)) withAutoprint(\{ # examplesIf}
w <- createWeights(data = data, formulas = f,
                   method = "bart")
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("SuperLearner", quietly = TRUE)) withAutoprint(\{ # examplesIf}
w <- createWeights(data = data, formulas = f,
                   method = "super")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[WeightIt:weightitMSM]{WeightIt::weightitMSM()}}
}
