% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCovariates.R
\name{dataToImpute}
\alias{dataToImpute}
\title{Creates data structure for imputation}
\usage{
dataToImpute(object, all_potential_covariates)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{covariates_to_include}{output from identifyPotentialConfounds}
}
\value{
data_to_impute
}
\description{
Creates data structure for imputation with only the variables needed (i.e., exposure, outcome, potential confounds)
}
\examples{
dataToImpute(object, all_potential_covariates)

}
\seealso{
\code{\link[=formatDataStruct]{formatDataStruct()}}, \code{\link[=identifyPotenialConfounds]{identifyPotenialConfounds()}}
}
