% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCausalQuestion.R
\name{defineCausalQuestion}
\alias{defineCausalQuestion}
\title{Define causal questions at the heart of the subsequent marginal structural models}
\usage{
defineCausalQuestion(object)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}
}
\description{
This code helps the individual delineate core aspects of their causal questions
}
\examples{
defineCausalQuestion(object)

}
