% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devMSMs-package.R
\name{devMSM_common_docs}
\alias{devMSM_common_docs}
\title{Common documentation for devMSMs functions}
\usage{
devMSM_common_docs(
  data,
  obj,
  formulas,
  weights,
  fit,
  bal_stats,
  verbose,
  save.out,
  home_dir,
  i,
  t
)
}
\arguments{
\item{data}{data in wide format as: a data frame, list of imputed data
frames, or \code{mids} object from the \code{mice} package}

\item{obj}{initialized MSM object from \code{\link[=initMSM]{initMSM()}}}

\item{formulas}{list of balancing formulas at each time point output from \code{\link[=createFormulas]{createFormulas()}}}

\item{weights}{list of IPTW weights output from \code{\link[=createWeights]{createWeights()}}}

\item{fit}{list of model outputs from \code{\link[=fitModel]{fitModel()}}}

\item{bal_stats}{list of balance statistics from \code{\link[=assessBalance]{assessBalance()}}}

\item{verbose}{(optional) TRUE or FALSE indicator for printing output to console
(default is FALSE)}

\item{save.out}{(optional) TRUE or FALSE indicator to save output and
intermediary output locally (default is FALSE)}

\item{home_dir}{path to home directory (required if save.out = TRUE)}

\item{i}{For multiply imputed datset, \code{i} selects which imputation to print results for.
Default is \code{i = 1}. With \code{i = TRUE}, all imputed datasets will be looped over.}

\item{t}{Which exposure variable to use. Can either be an index from 1, ..., num exposures or
a string containing the exposure varaible name (e.g. \code{"A.3"}). With \code{t = TRUE}, all exposure
variables will be looped over.}
}
\description{
Common documentation for devMSMs functions
}
\keyword{internal}
