% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.add_histories}
\alias{.add_histories}
\title{Add history labels and dose to table based on \code{epoch_vars}}
\usage{
.add_histories(p, epoch_vars)
}
\arguments{
\item{p}{table output from marginaleffects::avg_predictions() or hypotheses()}

\item{epoch_vars}{variable names of exposure epoch (in time order)}
}
\value{
table with histories labeled
}
\description{
Add history labels and dose to table based on \code{epoch_vars}
}
\keyword{internal}
