% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.avg_imp_bal_stats}
\alias{.avg_imp_bal_stats}
\title{Added by IS: averages balance stats across time points and recalculates balance:
returns one bal stat per covar}
\usage{
.avg_imp_bal_stats(bal_stats)
}
\description{
Added by IS: averages balance stats across time points and recalculates balance:
returns one bal stat per covar
}
\keyword{internal}
