% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.avg_imp_bal_stats_time}
\alias{.avg_imp_bal_stats_time}
\title{Added by IS: Averages balance stats across imputed datasets: list by time}
\usage{
.avg_imp_bal_stats_time(bal_stats)
}
\description{
Added by IS: Averages balance stats across imputed datasets: list by time
}
\keyword{internal}
