% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.characterize_exposure}
\alias{.characterize_exposure}
\title{Used in \code{assessBalance} to define history}
\usage{
.characterize_exposure(mat, exposure_type)
}
\description{
Used in \code{assessBalance} to define history
}
\keyword{internal}
