% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.create_hypotheses_mat}
\alias{.create_hypotheses_mat}
\title{Create contrast map using comparison and reference}
\usage{
.create_hypotheses_mat(histories, reference, comparison)
}
\arguments{
\item{histories}{character vector (from preds$term corresponding to average predictions)}

\item{reference}{character vector of exposure histories}

\item{comparison}{character vector of exposure histories}
}
\description{
Create contrast map using comparison and reference
}
\keyword{internal}
