% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.dose_from_history}
\alias{.dose_from_history}
\title{Calculate number of doses from history string (e.g. "l-l-l")}
\usage{
.dose_from_history(history, dose_level = "h")
}
\description{
Calculate number of doses from history string (e.g. "l-l-l")
}
\examples{
library(devMSMs)
devMSMs:::.dose_from_history(c("l-l-l-h", "l-l-l-l"), dose_level = "h")
devMSMs:::.dose_from_history(c("l-l-l-h", "l-l-l-l"), dose_level = "l")

}
\keyword{internal}
