% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.extract_time_pts_from_vars}
\alias{.extract_time_pts_from_vars}
\title{Grabs trailing numbers after the last occurance of \code{sep} in \code{vars}}
\usage{
.extract_time_pts_from_vars(vars, sep = "[\\\\._]")
}
\description{
Grabs trailing numbers after the last occurance of \code{sep} in \code{vars}
}
\keyword{internal}
