% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.generate_epoch_exposures}
\alias{.generate_epoch_exposures}
\title{Takes row means of eposure variables in the same epoch}
\usage{
.generate_epoch_exposures(data, exposure, epoch, sep)
}
\description{
Takes row means of eposure variables in the same epoch
}
\keyword{internal}
