% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_avg_predictions_variables}
\alias{.get_avg_predictions_variables}
\title{Grab values of epoch variables to evaluate in \code{marginaleffects::predictions}}
\usage{
.get_avg_predictions_variables(
  data,
  epoch_vars,
  exposure_type,
  hi_lo_cut = NULL
)
}
\description{
For continuous variables, grab quantiles of the epoch variables based on hi_lo_cut
For binary variables, use 0 and 1.
}
\keyword{internal}
