% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_epoch_var_names}
\alias{.get_epoch_var_names}
\title{Simple function to get the variable names associated with epochs}
\usage{
.get_epoch_var_names(exposure, epoch, sep)
}
\description{
Simple function to get the variable names associated with epochs
}
\keyword{internal}
