% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_history_values}
\alias{.get_history_values}
\title{Finds custom reference values (currently unused)}
\usage{
.get_history_values(epoch_d, exp_history)
}
\arguments{
\item{epoch_d}{data frame of high and low values per exposure main effect}

\item{exp_history}{sequence of "h" and/or "l" (e.g., "h-h-h")}
}
\value{
exposure history values
}
\description{
Finds custom reference values (currently unused)
}
\examples{
epoch_d <- data.frame(
  e = c("A.1", "A.2", "A.3"),
  l = c(0, 0, 0),
  h = c(1, 1, 1)
)
r <- devMSMs:::.get_history_values(
  epoch_d = epoch_d,
  exp_history = "l-l-l"
)
r <- devMSMs:::.get_history_values(
  epoch_d = epoch_d,
  exp_history = "h-h-h"
)

}
\keyword{internal}
