% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_interaction_terms}
\alias{.get_interaction_terms}
\title{Return interactions up to order n of variables in \code{x}}
\usage{
.get_interaction_terms(x, n = 2)
}
\arguments{
\item{x}{character vector of variable names}

\item{n}{order of interaction}
}
\description{
Return interactions up to order n of variables in \code{x}
}
\keyword{internal}
