% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.remove_time_pts_from_vars}
\alias{.remove_time_pts_from_vars}
\title{Gets prefix of variable by numbers after last occurance of \code{sep} and optionally deleting \code{sep}}
\usage{
.remove_time_pts_from_vars(vars, sep = "[\\\\._]", keep_sep = FALSE)
}
\description{
Gets prefix of variable by numbers after last occurance of \code{sep} and optionally deleting \code{sep}
}
\keyword{internal}
