% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.string_count}
\alias{.string_count}
\title{Base R rewrite of stringr::str_count(); pattern can only be length 1}
\usage{
.string_count(string, pattern = "")
}
\description{
Base R rewrite of stringr::str_count(); pattern can only be length 1
}
\keyword{internal}
