% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_hist.R
\name{eval_hist}
\alias{eval_hist}
\title{Visualize distribution of sample across exposure histories}
\usage{
eval_hist(
  data,
  exposure,
  tv_confounders,
  epochs = NULL,
  time_pts,
  hi_lo_cut = NULL,
  ref = NA,
  comps = NULL,
  verbose = TRUE
)
}
\arguments{
\item{exposure}{name of exposure variable}

\item{tv_confounders}{list of time-varying confounders with ".timepoint"
suffix}

\item{hi_lo_cut}{list of two numbers indicating quantile values that reflect
high and low values, respectively, for continuous exposure}

\item{verbose}{(optional) TRUE or FALSE indicator for user output (default is
TRUE)}
}
\value{
none
}
\description{
Create customized, user-specified exposure histories and tables displaying
sample distribution across them for user inspection.
}
