% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessDataStruct.R
\name{formatDataStruct}
\alias{formatDataStruct}
\title{Formats dataset and creates required directory}
\usage{
formatDataStruct(msm_object)
}
\arguments{
\item{msm_object}{msm object that contains all relevant user inputs}
}
\value{
data formatted dataset
}
\description{
Creates the required directories, assesses the data structure, and makes it uniform for future functions and returns data
This function requires a clean dataset in long format that contains columns for ID, time, exposure, outcome, and potential covariate confounds
}
\examples{
formatDataStruct(object)
}
