% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatForWeights.R
\name{formatForWeights}
\alias{formatForWeights}
\title{formatForWeights}
\usage{
formatForWeights(
  ID,
  home_dir,
  m,
  data,
  imputed_datasets = list(),
  time_varying_covariates,
  time_pts,
  time_var_exclude = NULL,
  just_imputed = "yes"
)
}
\arguments{
\item{ID}{identifier}

\item{home_dir}{path to home directory for project}

\item{m}{number of imputed datasets from Amelia}

\item{data}{output from formatDataStruct}

\item{imputed_datasets}{output from imputeData}

\item{time_pts}{identifier}

\item{time_var_exclude}{list any time-varying variables that should not be present because of planned missingness design}

\item{just_imputed}{"yes"= you have imputed datasets in global environment or "no" but they are saved locally from previous run}
}
\value{
wide_long_datasets
}
\description{
formatForWeights
}
