% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatLongData.R
\name{formatLongData}
\alias{formatLongData}
\title{Formats long data}
\usage{
formatLongData(
  home_dir,
  data,
  exposure,
  exposure_time_pts,
  outcome,
  tv_confounders,
  time_var = NA,
  id_var = NA,
  missing = NA,
  factor_confounders = NULL
)
}
\arguments{
\item{home_dir}{path to home directory}

\item{data}{dataframe in long format}

\item{exposure}{name of exposure variable}

\item{exposure_time_pts}{list of integers at which weights will be
created/assessed that correspond to time points when exposure was measured}

\item{outcome}{name of outcome variable with ".timepoint" suffix}

\item{tv_confounders}{list of time-varying confounders with ".timepoint"
suffix}

\item{time_var}{(optional) variable name in original dataset demarcating time}

\item{id_var}{(optional) variable name in original dataset demarcating ID}

\item{missing}{(optional) indicator for missing data in original dataset}

\item{factor_confounders}{(optional) list of variable names that are factors
(default is numeric)}
}
\value{
formatted long dataset
}
\description{
Formats long data
}
