% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHelpers.R
\name{get_comparison_values}
\alias{get_comparison_values}
\title{Finds custom comparison values}
\usage{
get_comparison_values(d, comp_histories)
}
\arguments{
\item{comp_histories}{}
}
\value{
comparison values
}
\description{
Finds custom comparison values
}
