% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModel.R
\name{getModel}
\alias{getModel}
\title{Fits outcome model}
\usage{
getModel(
  d,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  interactions,
  verbose
)
}
\arguments{
\item{d}{wide data frame}

\item{outcome}{name of outcome variable with ".timepoint" suffix}

\item{epochs}{data frame of exposure epoch labels and values}

\item{int_order}{integer specification of highest order exposure main effects interaction for interaction models}

\item{fam}{function specification for svyglm model}

\item{covariates}{list of characters reflecting variable names of covariates for covariate models}

\item{verbose}{TRUE or FALSE indicator for user output}
}
\value{
list of fitted model(s)
}
\description{
Fits outcome model
}
