% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessBalance.R
\name{getWeights}
\alias{getWeights}
\title{Get weights
Code to get and assign weights output if code has already been run to save weights locally}
\usage{
getWeights(home_dir, just_made_weights = "no")
}
\arguments{
\item{home_dir}{path to home directory for the project}

\item{just_made_weights="no"}{no=weights are not assigned in global environment and are instead saved locally}
}
\value{
weights_models
}
\description{
Get weights
Code to get and assign weights output if code has already been run to save weights locally
}
\examples{
getWeights(home_dir, just_made_weights="no")
}
\seealso{
\code{\link[=assessBalance]{assessBalance()}} for more on the weights_models param
}
