% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCovariates.R
\name{identifyPotentialConfounds}
\alias{identifyPotentialConfounds}
\title{Identify potential confounds}
\usage{
identifyPotentialConfounds(object, time_pt_datasets)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{time_pt_datasets}{output from makeTimePtDatasets}
}
\value{
covariates_to_include
}
\description{
Identifies potential confounds from the covariates based on those that are correlated with either exposures or outcomes above 0.1
}
\examples{
identifyPotentialConfounds(home_dir, data, time_pt_datasets, time_pts, exclude_covariates=NULL)

}
\seealso{
\code{\link[=formatDataStruct]{formatDataStruct()}}, \code{\link[=makeTimePtDatasets]{makeTimePtDatasets()}}
}
