% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeData.R
\name{imputeData}
\alias{imputeData}
\title{Imputes dataset so there is no missing at each time point using parallel processing to speed up}
\usage{
imputeData(object, data_to_impute, read_imps_from_file = "no")
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{data_to_impute}{output from dataToImpute}
}
\value{
imputed_datasets imputation results
}
\description{
Creates m imputed datasets from original datasets using mice::mice
}
\examples{
imputeData(object, data_to_impute, read_imps_from_file="no")

}
\seealso{
\code{\link[=dataToImpute]{dataToImpute()}}, \code{\link[mice:mice]{mice::mice()}}
}
