% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_love_plot.R
\name{make_love_plot}
\alias{make_love_plot}
\title{Create love plots showing balancing statistics}
\usage{
make_love_plot(
  home_dir,
  folder,
  exposure,
  exposure_time_pt,
  exposure_type,
  k = 0,
  form_name,
  balance_stats,
  data_type,
  balance_thresh,
  weights_method,
  imp_conf,
  verbose,
  save.out
)
}
\arguments{
\item{home_dir}{path to home directory}

\item{folder}{folder path for saving}

\item{exposure}{name of exposure variable}

\item{exposure_time_pt}{exposure time point integer}

\item{exposure_type}{character indicating binary or continuous exposure type}

\item{k}{imputation number}

\item{form_name}{formula name}

\item{balance_stats}{data frame of balance statistics}

\item{balance_thresh}{one or two numbers between 0 and 1 indicating a single balancing threshold or thresholds for more and less important confounders, respectively}

\item{weights_method}{method character string of WeightItMSM() balancing method abbreviation}

\item{imp_conf}{list of variable names reflecting important confounders}

\item{save.out}{TRUE or FALSE indicator to save output and intermediary output locally}
}
\value{
none
}
\description{
Create love plots showing balancing statistics
}
