% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessDataStruct.R
\name{makeTimePtDatasets}
\alias{makeTimePtDatasets}
\title{Makes timepoint datasets}
\usage{
makeTimePtDatasets(data, ID, time_pts)
}
\arguments{
\item{data}{output from formatDataStruct}

\item{ID}{person-level identifier in your dataset}

\item{time_pts}{list of time points along your developmental path of interest for which you have at least one measurement}
}
\value{
time_pt_datasets datasets split by time point
}
\description{
Creates datasets unique to each time point for future use and returns a list of time point datasets
}
\examples{
makeTimePtDatasets(data, ID, time_pts)

}
\seealso{
\code{\link[=fomatDataStruct]{fomatDataStruct()}}
}
