% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcCorrection.R
\name{mcCorrection}
\alias{mcCorrection}
\title{Multiple comparison correction
Adjusts p-values for multiple comparisons by exposure-outcome pairing}
\usage{
mcCorrection(history_comparisons, method = "BH")
}
\arguments{
\item{history_comparisons}{output from compareHistories}

\item{method}{abbreviation for multiple comparison correction method (see p.adjust)}
}
\value{
significant_comparisons
}
\description{
Multiple comparison correction
Adjusts p-values for multiple comparisons by exposure-outcome pairing
}
\examples{
mcCorrection(history_comparisons, method="BH")
}
\seealso{
\code{\link[stats:p.adjust]{stats::p.adjust()}} \code{\link[=comparisonHistories]{comparisonHistories()}}
}
