% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmObject.R
\name{msmObject}
\alias{msmObject}
\title{Create msm object
Gathers all information needed for the functions}
\usage{
msmObject(
  data_path,
  home_dir,
  ID,
  time_pts,
  time_var,
  missing,
  m,
  exposures,
  exposure_time_pts,
  exposure_epochs,
  outcomes,
  outcome_time_pts,
  continuous_variables = NULL,
  time_var_exclude = NULL
)
}
\arguments{
\item{data_path}{path to cleaned dataset}

\item{home_dir}{path to home directory for the project}

\item{ID}{person-level identifier in your dataset}

\item{time_pts}{list of time points along your developmental path of interest for which you have at least one measurement}

\item{time_var}{variable in your long dataset that designates developmental time}

\item{missing}{missing data marker in your dataset}

\item{m}{number of imputations}

\item{exposures}{list of variables that represent your exposures/treatments of interest}

\item{exposure_time_pts}{times when exposures occurred}

\item{exposure_epochs}{data frame containing epochs and correponding time points for exposure histories}

\item{outcomes}{list of variables that represent your outcomes of interest}

\item{outcome_time_pts}{list of time points for outcomes}

\item{continuous_variables}{list of continuous variables in your dataset}

\item{time_var_exclude}{time-varying variables that should not be present because of planned missingness}
}
\description{
Create msm object
Gathers all information needed for the functions
}
\examples{
msmObject(data_path, home_dir, ID, time_pts, time_var, missing, m, exposures, exposure_time_pts, exposure_epochs, outcomes, outcome_time_pts, continuous_variables, time_var_exclude)
}
