% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHelpers.R
\name{perform_multiple_comparison_correction}
\alias{perform_multiple_comparison_correction}
\title{Conduct multiple comparison correction}
\usage{
perform_multiple_comparison_correction(
  comps,
  reference,
  comp_histories,
  method,
  verbose
)
}
\arguments{
\item{comps}{table}

\item{reference}{reference sequence of "h" and/or "l" (e.g., "h-h-h")}

\item{comp_histories}{comparison sequence(s) of "h" and/or "l" (e.g., "h-h-h")}

\item{method}{character abbreviation for multiple comparison correction method}

\item{verbose}{(optional) TRUE or FALSE indicator for user output (default is
TRUE)}
}
\value{
comparison table with corrected p-values
}
\description{
Conduct multiple comparison correction
}
