% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perm2}
\alias{perm2}
\title{Adaptation of gtools::permutations(2, repeats.allowed = TRUE)}
\usage{
perm2(r, v)
}
\description{
Adaptation of gtools::permutations(2, repeats.allowed = TRUE)
}
\keyword{internal}
