% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotResults}
\alias{plotResults}
\title{Plot results from history comparisons
Code to plot predicted values of the different exposure histories by history and colored by dosage of exposure}
\usage{
plotResults(object, history_estimates)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{best_models}{outcome from assessModels}
}
\value{
plots
}
\description{
Plot results from history comparisons
Code to plot predicted values of the different exposure histories by history and colored by dosage of exposure
}
\examples{
plotResults(object, best_models)
}
\seealso{
\code{\link[=assessModels]{assessModels()}}
}
