% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preBalanceChecking.R
\name{preBalanceChecking}
\alias{preBalanceChecking}
\title{Examining initial imbalance for each exposure, each time point, each covariate for each history using the Jackson method}
\usage{
preBalanceChecking(object, wide_long_datasets, all_forms, histories = 1)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{wide_long_datasets}{from formatForWeights()}

\item{all_forms}{from createForms()}

\item{histories}{binary indicator of whether to print history sample distributions}
}
\description{
Examining initial imbalance for each exposure, each time point, each covariate for each history using the Jackson method
}
\examples{
preBalanceChecking(object, wide_long_datasets, all_forms, histories=1)

}
