% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_eval_hist.R
\name{print_eval_hist}
\alias{print_eval_hist}
\title{Visualize distribution of sample across exposure histories}
\usage{
print_eval_hist(
  epoch_history,
  epoch,
  hi_lo_cut,
  reference = NULL,
  comparison = NULL
)
}
\arguments{
\item{epoch_history}{character vector of length \code{nrow(data)} containing epochi histories for
each unit}

\item{epoch}{character vector of epoch values}

\item{hi_lo_cut}{list of two numbers indicating quantile values
that reflect high and low values, respectively, for continuous exposure
(default is median split)}

\item{reference}{lists of one or more strings of "-"-separated "l"
and "h" values indicative of a reference exposure history to which to
compare comparison, required if comparison is supplied}

\item{comparison}{(optional) list of one or more strings of "-"-separated "l"
and "h" values indicative of comparison history/histories to compare to
reference, required if reference is supplied}
}
\value{
invisibly NULL
}
\description{
Create customized, user-specified exposure histories and tables displaying
sample distribution across them for user inspection.
}
\keyword{internal}
