% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data_wide_bin.rda}
\alias{sim_data_wide_bin.rda}
\alias{sim_data_wide_bin}
\title{Wide complete data (binary exposure)
These data are simulated based on data from the Family Life Project (FLP), a longitudinal study following 1,292 families
representative of two geographic areas (three counties in North Carolina and three counties in Pennsylvania) with high rural
child poverty (Vernon-Feagans et al., 2013; Burchinal et al., 2008). These data contain economic strain (ESEATA1) as a binary variable.}
\format{
A data frame
}
\usage{
sim_data_wide_bin
}
\description{
Wide complete data (binary exposure)
These data are simulated based on data from the Family Life Project (FLP), a longitudinal study following 1,292 families
representative of two geographic areas (three counties in North Carolina and three counties in Pennsylvania) with high rural
child poverty (Vernon-Feagans et al., 2013; Burchinal et al., 2008). These data contain economic strain (ESEATA1) as a binary variable.
}
\keyword{datasets}
