% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimWeights.R
\name{trimWeights}
\alias{trimWeights}
\title{Trim IPTW balancing weights}
\usage{
trimWeights(home_dir, weights, quantile = 0.95, verbose = TRUE)
}
\arguments{
\item{home_dir}{path to home directory}

\item{weights}{list of IPTW weights output from createWeights()}

\item{quantile}{numeric value between 0 and 1 of quantile value at which to trim weights}

\item{verbose}{TRUE or FALSE indicator for user output}
}
\value{
list of model output with trimmed weights
}
\description{
Trim IPTW balancing weights
}
