% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateWeights.R
\name{truncateWeights}
\alias{truncateWeights}
\title{Truncates weights}
\usage{
truncateWeights(object, wide_long_datasets, imp_data_w)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{data_for_model_with_weights}{output from condenseWeights}
}
\value{
data_for_model_with_weights_cutoff
}
\description{
Code to 'cut off' weights at 90th percentile and populate all of those above at 90th percentile value to avoid heavy tails that can bias results
}
\examples{
truncateWeights(data_for_model_with_weights, home_dir, exposures)

}
\seealso{
\code{\link[=condenseWeights]{condenseWeights()}}
}
