% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateForms.R
\name{updateForms}
\alias{updateForms}
\title{Creating updated short forms with any imbalanced covariates at time lags greater than the user-specified lag}
\usage{
updateForms(msm_object, forms, data_for_model_with_weights, balance_stats_full)
}
\arguments{
\item{msm_object}{msm object that contains all relevant user inputs}

\item{forms}{short forms from createShortForms}

\item{data_for_model_with_weights}{all imputed datasets with weights}

\item{balance_stats_full}{balance assessment using full forms}
}
\value{
new_forms
}
\description{
Creating updated short forms with any imbalanced covariates at time lags greater than the user-specified lag
}
\examples{
updateForms(object, forms, data_for_model_with_weights, balance_stats_full)
}
