% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.inp.file.R
\name{write.inp.file}
\alias{write.inp.file}
\title{Title
This code comes from the aMNLFA package by Gottfredson and colleagues (2019)}
\usage{
write.inp.file(df, outfile)
}
\arguments{
\item{df}{dataframe}

\item{outfile}{path to output file}
}
\description{
Title
This code comes from the aMNLFA package by Gottfredson and colleagues (2019)
}
\examples{
write.inp.file(df, outfile)
}
\seealso{
\code{\link[aMNLFA:aMNLFA.initial]{aMNLFA::aMNLFA.initial()}}
}
