% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHelpers.R
\name{add_histories}
\alias{add_histories}
\title{Add history labels to table}
\usage{
add_histories(p, d)
}
\arguments{
\item{p}{table output from marginaleffects::avg_predictions() or hypotheses()}

\item{d}{data frame of high and low values per exposure main effect}
}
\value{
table with histories labeled
}
\description{
Add history labels to table
}
