% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessBalance.R
\name{assessBalance}
\alias{assessBalance}
\title{Assess success of covariate balancing}
\usage{
assessBalance(
  home_dir,
  data,
  exposure,
  exposure_time_pts,
  outcome,
  tv_confounders,
  type,
  formulas,
  weights = NULL,
  balance_thresh = 0.1,
  imp_conf = NULL,
  user.o = TRUE
)
}
\description{
Assess success of covariate balancing
}
