% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessBalance.R
\name{assessBalance}
\alias{assessBalance}
\title{Assess success of covariate balancing}
\usage{
assessBalance(
  home_dir,
  weights_models = list(),
  m,
  exposures,
  outcomes,
  time_pts,
  balance_thresh = 0.12,
  just_made_weights = "no"
)
}
\arguments{
\item{home_dir}{path to home directory for the project}

\item{weights_models}{list of weights models from CBPS}

\item{m}{number of imputations from Amelia}

\item{exposures}{list of variables that represent your exposures/treatments of interest}

\item{time_pts}{list of time points along your developmental path of interest for which you have at least one measurement}

\item{balance_thresh}{correlation value between covariate and exposure over which it is not considered balanced}

\item{just_made_weights="no"}{no=weights are not assigned in global environment and are instead saved locally}
}
\value{
list of unbalanced_covariates_for_models for each exposure
}
\description{
#Assesses how well balance was achieved for each of the covariates/potential confounds in relation to each of the exposures,
and returns a list of unbalanced covariates for each exposure to add to future models
}
\examples{
assessBalance(home_dir, weights_models=list(), m, exposures, time_pts, balance_thresh=0.12, just_made_weights="no")
}
\seealso{
\code{\link[CBPS:balance]{CBPS::balance()}} for more on the balance function

\code{\link[=createWeights]{createWeights()}} for more on the weights_models param
}
