% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessBalance.R
\name{assessBalance}
\alias{assessBalance}
\title{Assesses confounder balancing}
\usage{
assessBalance(
  data,
  obj,
  weights = NULL,
  balance_thresh = NULL,
  imp_conf = NULL,
  verbose = FALSE,
  save.out = FALSE,
  home_dir = NULL
)
}
\arguments{
\item{data}{data in wide format as: a data frame, list of imputed data
frames, or mids object}

\item{obj}{initialized MSM object from \code{initMSM}}

\item{weights}{list of IPTW weights output from \code{createWeights},
required for type 'weighted'}

\item{balance_thresh}{(optional) one or two numbers between 0 and 1
indicating a single balancing threshold or thresholds for more and less
important confounders, respectively (default = 0.1)}

\item{imp_conf}{(optional) list of variable names reflecting important
confounders, required if two balance thresholds are supplied}

\item{verbose}{(optional) TRUE or FALSE indicator for user output (default is
TRUE)}

\item{save.out}{(optional) TRUE or FALSE indicator to save output and
intermediary output locally (default is TRUE)}

\item{home_dir}{(optional) path to home directory
(required if save.out = TRUE)}
}
\value{
a data frame of balance statistics
}
\description{
Draws on functions from the cobalt package to quantify the relations between
exposure and confounders at each exposure time point according to the
guidelines from Jackson, 2016 on how to assess balance for time-varying
exposures.
}
\examples{
library(devMSMs)
data <- data.frame(
  ID = 1:50,
  A.1 = rnorm(n = 50),
  A.2 = rnorm(n = 50),
  A.3 = rnorm(n = 50),
  B.1 = rnorm(n = 50),
  B.2 = rnorm(n = 50),
  B.3 = rnorm(n = 50),
  C = rnorm(n = 50),
  D.3 = rnorm(n = 50)
)
obj <- initMSM(
  data,
  exposure = c("A.1", "A.2", "A.3"),
  ti_conf = c("C"),
  tv_conf = c("B.1", "B.2", "B.3", "D.3")
)
f <- createFormulas(obj, type = "short")

# Prebalance
b <- assessBalance(data = data, obj = obj)
print(b)

# returns list of ggplots for each exposure variable
plots <- plot(b) 
plots[[1]]
plots[[2]]
plots[[3]]


# Weighted
w <- createWeights(data = data, obj = obj, formulas = f)
bw <- assessBalance(data = data, obj = obj, weights = w)
print(bw)

}
\seealso{
{\link{cobalt} package,
\url{https://cran.r-project.org/web/packages/cobalt/index.html}}

{Jackson, 2016 for more on assessing balance for time-varying
exposures, \url{https://pubmed.ncbi.nlm.nih.gov/27479649/}}
}
