% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessBalance.R
\name{assessBalance}
\alias{assessBalance}
\alias{print.devMSM_bal_stats}
\alias{summary.devMSM_bal_stats}
\alias{plot.devMSM_bal_stats}
\title{Assesses confounder balancing}
\usage{
assessBalance(
  data,
  obj,
  weights = NULL,
  balance_thresh = NULL,
  imp_conf = NULL,
  verbose = FALSE,
  save.out = FALSE
)

\method{print}{devMSM_bal_stats}(x, i = NA, t = TRUE, ...)

\method{summary}{devMSM_bal_stats}(object, i = NA, t = TRUE, ...)

\method{plot}{devMSM_bal_stats}(x, i = NA, t = TRUE, ...)
}
\arguments{
\item{data}{data in wide format as: a data frame, list of imputed data
frames, or \code{mids} object from the \code{mice} package}

\item{obj}{initialized MSM object from \code{\link[=initMSM]{initMSM()}}}

\item{weights}{(optional) list of IPTW weights output from \code{\link[=createWeights]{createWeights()}}}

\item{balance_thresh}{(optional) one or two numbers between 0 and 1
indicating a single balancing threshold or thresholds for more and less
important confounders, respectively (default = 0.1)}

\item{imp_conf}{(optional) list of variable names reflecting important
confounders, required if two balance thresholds are supplied}

\item{verbose}{(optional) TRUE or FALSE indicator for printing output to console
(default is FALSE)}

\item{save.out}{(optional) TRUE or FALSE indicator to save output and
intermediary output locally (default is FALSE)}

\item{x}{devMSM_bal_stats object from \code{assessBalance}}

\item{i}{For multiply imputed datset, \code{i} selects which imputation to print results for.
Default is \code{i = 1}. With \code{i = TRUE}, all imputed datasets will be looped over.}

\item{t}{Which exposure variable to use. Can either be an index from 1, ..., num exposures or
a string containing the exposure varaible name (e.g. \code{"A.3"}). With \code{t = TRUE}, all exposure
variables will be looped over.}

\item{...}{ignored}

\item{object}{devMSM_bal_stats object from \code{\link[=assessBalance]{assessBalance()}}}
}
\value{
a list containing balance statistics as a dataframe. It is the length
of the number of datasets (1 for a data.frame or the number of imputed datasets)
}
\description{
Draws on functions from the cobalt package to quantify the relations between
exposure and confounders at each exposure time point according to the
guidelines from Jackson, 2016 on how to assess balance for time-varying
exposures.
}
\examples{
library(devMSMs)
data <- data.frame(
  ID = 1:50,
  A.1 = rnorm(n = 50),
  A.2 = rnorm(n = 50),
  A.3 = rnorm(n = 50),
  B.1 = rnorm(n = 50),
  B.2 = rnorm(n = 50),
  B.3 = rnorm(n = 50),
  C = rnorm(n = 50),
  D.3 = rnorm(n = 50)
)
obj <- initMSM(
  data,
  exposure = c("A.1", "A.2", "A.3"),
  ti_conf = c("C"),
  tv_conf = c("B.1", "B.2", "B.3", "D.3")
)
f <- createFormulas(obj, type = "short")

# Prebalance
b <- assessBalance(data = data, obj = obj)
print(b)

# returns ggplot of balance stats for all exposure variables
plots <- plot(b, t = TRUE)
# can plot only specific exposure time periods
plot(b, t = "A.3")
plot(b, t = 3)

# Weighted
w <- createWeights(data = data, obj = obj, formulas = f)
bw <- assessBalance(data = data, obj = obj, weights = w)
print(bw)
plot(bw)

}
\seealso{
\link{cobalt} package, \url{https://cran.r-project.org/web/packages/cobalt/index.html};
Jackson, 2016 for more on assessing balance for time-varying exposures,
\url{https://pubmed.ncbi.nlm.nih.gov/27479649/}
}
