% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessBalance.R
\name{assessBalance}
\alias{assessBalance}
\title{Assess success of covariate balancing
Assesses how well balance was achieved for each of the covariates/potential confounds in relation to each of the exposure,
and returns a list of unbalanced covariates for each exposure to add to future models}
\usage{
assessBalance(
  home_dir,
  data,
  exposure,
  outcome,
  tv_confounders,
  type,
  formulas,
  weights = NULL,
  balance_thresh = 0.1,
  imp_conf = NULL,
  user.o = T
)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{forms}{formula for assessing balance}

\item{data_for_model_with_weights}{imputed data with weights}

\item{histories}{optional binary indicator of whether to print histories for bal stats}
}
\value{
list of unbalanced_covariates_for_models for each exposure
}
\description{
Assess success of covariate balancing
Assesses how well balance was achieved for each of the covariates/potential confounds in relation to each of the exposure,
and returns a list of unbalanced covariates for each exposure to add to future models
}
\examples{
assessBalance(object, forms, data_for_model_with_weights, histories=1)
}
\seealso{
\code{\link[=msmObject]{msmObject()}} for more on the weights_models param

\code{\link[=createWeights]{createWeights()}} for more on the weights_models param
}
