% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessModels.R
\name{assessModel}
\alias{assessModel}
\title{Assess marginal structural models
Code to assess the various models and decide which fits the best}
\usage{
assessModel(object, all_models, best_model = NULL)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{all_models}{output from fitModel}
}
\value{
best_models list of best-fitting models for each exposure=outcome pairing
}
\description{
Assess marginal structural models
Code to assess the various models and decide which fits the best
}
\examples{
assessModel(object, all_models)
}
\seealso{
\code{\link[=fitModel]{fitModel()}}
}
