% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBalStats.R
\name{calcBalStats}
\alias{calcBalStats}
\title{Calculate balance stats based on Jackson paper}
\usage{
calcBalStats(data, obj, weights = NULL, balance_thresh = NULL, imp_conf = NULL)
}
\arguments{
\item{data}{data in wide format as: a data frame, path to folder of imputed
.csv files, or mids object}

\item{obj}{initialized MSM object from \code{initMSM}}

\item{weights}{(optional) list of IPTW weights output from createWeights}

\item{balance_thresh}{(optional) one or two numbers between 0 and 1
indicating a single balancingn threshold or thresholds for more and less
important confounders, respectively}

\item{imp_conf}{(optional) list of variable names reflecting important
confounders (required if two balance thresholds are provided)}
}
\value{
data frame of balance statistics
}
\description{
Calculate weighted or unweighted standardized balance statistics for a given
exposure time point, using all relevant confounders. Draws on Jackson, 2016
approaches to assessing balance for time-varying exposures by weighting
statistics based on sample distribution in exposure histories.
}
