% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBalStats.R
\name{calcBalStats}
\alias{calcBalStats}
\title{Code to calculate balance stats based on Jackson paper (either weighted or unweighted)}
\usage{
calcBalStats(
  data,
  formulas,
  exposure,
  exposure_time_pts,
  outcome,
  balance_thresh,
  k = 0,
  weights = NULL,
  imp_conf = NULL,
  user.o = TRUE
)
}
\description{
Code to calculate balance stats based on Jackson paper (either weighted or unweighted)
}
