% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHistories.R
\name{compareHistories}
\alias{compareHistories}
\title{Compare exposure histories
This code uses the best-fitting model for each exposure-outcome pair to compare the effects of user-specified reference and comparison histories of exposure on outcome u sing linear hypothesis testing}
\usage{
compareHistories(
  object,
  hi_cutoff = 0.75,
  lo_cutoff = 0.25,
  best_models,
  comparisons = "",
  reference = ""
)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{hi_cutoff}{integer for percentile considered "high" for exposure}

\item{lo_cutoff}{integer for percentile considered "low" for exposure}

\item{best_models}{best-fitting models outcome from assessModel}

\item{comparisons}{list of histories to compare to the reference history}

\item{reference}{reference history to compare comparison histories}
}
\value{
history_comparisons lists of linear hypothesis tests
}
\description{
Compare exposure histories
This code uses the best-fitting model for each exposure-outcome pair to compare the effects of user-specified reference and comparison histories of exposure on outcome u sing linear hypothesis testing
}
\examples{
compareHistories(object, hi_cutoff=.75, lo_cutoff=.25, best_models, comparisons, reference)
}
\seealso{
\code{\link[=assessModel]{assessModel()}}
}
