% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHistories.R
\name{compareHistories}
\alias{compareHistories}
\title{Estimate, compare, and visualize exposure histories}
\usage{
compareHistories(
  home_dir,
  exposure,
  exposure_time_pts,
  outcome,
  model,
  epochs = NULL,
  hi_lo_cut = NULL,
  reference = NULL,
  comparison = NULL,
  mc_comp_method = NA,
  dose_level = NA,
  exp_lab = NA,
  out_lab = NA,
  colors = NULL,
  verbose = TRUE,
  save.out = TRUE
)
}
\arguments{
\item{home_dir}{path to home directory (required if 'save.out' = TRUE)}

\item{exposure}{name of exposure variable}

\item{exposure_time_pts}{list of integers at which weights will be
created/assessed that correspond to time points when exposure wass measured}

\item{outcome}{name of outcome variable with ".timepoint" suffix}

\item{model}{list of model outputs from fitModel()}

\item{epochs}{(optional) data frame of exposure epoch labels and values}

\item{hi_lo_cut}{(optional) list of two numbers indicating quantile values
that reflect high and low values, respectively, for continuous exposure
(default is median split)}

\item{reference}{(optional) list sof one or more strings of "-"-separated "l"
and "h" values indicative of a reference exposure history to which to
compare comparison, required if comparison is supplied}

\item{comparison}{(optional) list of one or more strings of "-"-separated "l"
and "h" values indicative of comparison history/histories to compare to
reference, required if reference is supplied}

\item{mc_comp_method}{(optional) character abbreviation for multiple
comparison correction method for stats::p.adjust, default is
Benjamini-Hochburg ("BH")}

\item{dose_level}{(optional) "l" or "h" indicating whether low or high doses
should be tallied in tables and plots (default is high "h")}

\item{exp_lab}{(optional) character label for exposure variable in plots
(default is variable name)}

\item{out_lab}{(optional) character label for outcome variable in plots
(default is variable name)}

\item{colors}{(optional) character specifying Brewer palette or list of
colors (n(epochs)+1) for plotting (default is "Dark2" palette)}

\item{verbose}{(optional) TRUE or FALSE indicator for user output (default is
TRUE)}

\item{save.out}{(optional) TRUE or FALSE indicator to save output and
intermediary output locally (default is TRUE)}
}
\value{
data frame of history comparisons
}
\description{
Takes fitted model output to created predicted values for user-specified
histories (pooling for imputed data), before conducting contrast comparisons
(pooling for imputed data), correcting for multiple comparisons, and then
plotting results.
}
\examples{
print("tbd")

}
\seealso{
{\code{\link[marginaleffects:predictions]{marginaleffects::avg_predictions()}},
\url{https://cran.r-project.org/web/packages/marginaleffects/marginaleffects.pdf}}

{\code{\link[marginaleffects:hypotheses]{marginaleffects::hypotheses()}},
\url{https://cran.r-project.org/web/packages/marginaleffects/marginaleffects.pdf}}

{\code{\link[stats:p.adjust]{stats::p.adjust()}},
\url{https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/p.adjust}}
}
