% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHistories.R
\name{compareHistories}
\alias{compareHistories}
\title{Compare exposure histories
This code uses the best-fitting model for each exposure-outcome pair to compare the effects of user-specified reference and comparison histories of exposure on outcome using linear hypothesis testing}
\usage{
compareHistories(
  home_dir,
  data_for_model_with_weights_cutoff,
  all_models,
  reference = NA,
  compare = NA
)
}
\arguments{
\item{data_for_model_with_weights_cutoff}{imputed datasets with truncated weights}

\item{all_models}{fitted models for each imputed dataset}

\item{reference}{optional reference event for custom comparison}

\item{compare}{optional comparison event(s) for custom comparison}

\item{object}{msm object that contains all relevant user inputs}
}
\value{
preds_pool
}
\description{
Compare exposure histories
This code uses the best-fitting model for each exposure-outcome pair to compare the effects of user-specified reference and comparison histories of exposure on outcome using linear hypothesis testing
}
\examples{
compareHistories(object, data_for_model_with_weights_cutoff, all_models, reference=NA, compare=NA)
}
