% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHistories.R
\name{compareHistories}
\alias{compareHistories}
\title{Estimate, compare, and visualize exposure histories}
\usage{
compareHistories(
  home_dir,
  exposure,
  exposure_time_pts,
  outcome,
  tv_confounders,
  model,
  epochs = NULL,
  hi_lo_cut = NULL,
  reference = NA,
  comparison = NULL,
  mc_comp_method = "BH",
  dose_level = "h",
  exp_lab = NA,
  out_lab = NA,
  colors = "Dark2",
  verbose = TRUE,
  save.out = TRUE
)
}
\arguments{
\item{home_dir}{path to home directory}

\item{exposure}{name of exposure variable}

\item{exposure_time_pts}{list of integers at which weights will be
created/assessed that correspond to time points when exposure wass measured}

\item{outcome}{name of outcome variable with ".timepoint" suffix}

\item{tv_confounders}{list of time-varying confounders with ".timepoint"
suffix}

\item{model}{list of model outputs from fitModel()}

\item{epochs}{(optional) data frame of exposure epoch labels and values}

\item{hi_lo_cut}{(optional) list of two numbers indicating quantile values
that reflect high and low values, respectively, for continuous exposure
(default is median split)}

\item{reference}{(optional) string of "-"-separated "l" and "h" values
indicative of a reference exposure history to which to compare comparison,
required if comparison is supplied}

\item{comparison}{(optional) list of one or more strings of "-"-separated "l"
and "h" values indicative of comparison history/histories to compare to
reference, required if reference is supplied}

\item{mc_comp_method}{(optional) character abbreviation for multiple
comparison correction method for stats::p.adjust, default is
Benjamini-Hochburg ("BH")}

\item{dose_level}{(optional) "l" or "h" indicating whether low or high doses
should be tallied in tables and plots (default is high "h")}

\item{exp_lab}{(optional) character label for exposure variable in plots
(default is variable name)}

\item{out_lab}{(optional) character label for outcome variable in plots
(default is variable name)}

\item{colors}{(optional) character specifying Brewer palette or list of
colors (n(epochs)+1) for plotting (default is "Dark2" palette)}

\item{verbose}{(optional) TRUE or FALSE indicator for user output (default is
TRUE)}

\item{save.out}{(optional) TRUE or FALSE indicator to save output and
intermediary output locally (default is TRUE)}
}
\value{
data frame of history comparisons
}
\description{
Takes fitted model output to created predicted values for user-specified
histories (pooling for imputed data), before conducting contrast comparisons
(pooling for imputed data), correcting for multiple comparisons, and then
plotting results.
}
\examples{
f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                   formulas = f,
                   save.out = FALSE)

m <- fitModel(data = test,
              weights = w,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
              model = "m0",
              save.out = FALSE)

r <- compareHistories(exposure = "A",
                      exposure_time_pts = c(1, 2, 3),
                      outcome = "D.3",
                      tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                      model = m,
                      save.out = FALSE)
r <- compareHistories(exposure = "A",
                      exposure_time_pts = c(1, 2, 3),
                      outcome = "D.3",
                      tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                      model = m,
                      reference = "l-l-l",
                      comparison = "h-h-h",
                      save.out = FALSE)
r <- compareHistories(exposure = "A",
                      exposure_time_pts = c(1, 2, 3),
                      outcome = "D.3",
                      tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                      model = m,
                      reference = "l-l-l",
                      comparison = c("h-h-h", "h-l-l"),
                      save.out = FALSE)
}
\seealso{
{\code{\link[marginaleffects:predictions]{marginaleffects::avg_predictions()}}, <url1>}

{\code{\link[marginaleffects:hypotheses]{marginaleffects::hypotheses()}}, <url1>}

{\code{\link[stats:p.adjust]{stats::p.adjust()}}, <url1>}
}
