% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHistories.R
\name{compareHistories}
\alias{compareHistories}
\title{Compare exposure histories
This code uses the best-fitting model for each exposure-outcome pair to compare the effects of user-specified reference and comparison histories of exposure on outcome u sing linear hypothesis testing}
\usage{
compareHistories(
  object,
  data_for_model_with_weights_cutoff,
  all_models,
  imp_data_w_t
)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{best_models}{best-fitting models outcome from assessModel}
}
\value{
history_comparisons lists of linear hypothesis tests
}
\description{
Compare exposure histories
This code uses the best-fitting model for each exposure-outcome pair to compare the effects of user-specified reference and comparison histories of exposure on outcome u sing linear hypothesis testing
}
\examples{
compareHistories(object, best_models)
}
\seealso{
\code{\link[=assessModel]{assessModel()}}
}
