% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHistories.R
\name{compareHistories}
\alias{compareHistories}
\alias{print.devMSM_comparisons}
\alias{plot.devMSM_comparisons}
\alias{summary.devMSM_comparisons}
\title{Estimate, compare, and visualize exposure histories}
\usage{
compareHistories(
  fit,
  hi_lo_cut,
  dose_level = "h",
  reference = NULL,
  comparison = NULL,
  outcome_level = NULL,
  mc_comp_method = "BH",
  verbose = FALSE,
  save.out = FALSE
)

\method{print}{devMSM_comparisons}(x, save.out = FALSE, ...)

\method{plot}{devMSM_comparisons}(
  x,
  colors = "Dark2",
  exp_lab = NULL,
  out_lab = NULL,
  save.out = FALSE,
  ...
)

\method{summary}{devMSM_comparisons}(object, type = "comps", ...)
}
\arguments{
\item{fit}{list of model outputs from \code{\link[=fitModel]{fitModel()}}}

\item{hi_lo_cut}{list of two numbers indicating quantile values
that reflect high and low values, respectively, for continuous exposure}

\item{dose_level}{(optional) "l" or "h" indicating whether low or high doses
should be tallied in tables and plots (default is high "h")}

\item{reference}{lists of one or more strings of "-"-separated "l"
and "h" values indicative of a reference exposure history to which to
compare comparison, required if comparison is supplied}

\item{comparison}{(optional) list of one or more strings of "-"-separated "l"
and "h" values indicative of comparison history/histories to compare to
reference, required if reference is supplied}

\item{outcome_level}{levels of a factor outcome within which the user wishes
compare histories (default is all levels), history predictions and comparisons
will be made within each outcome level.}

\item{mc_comp_method}{(optional) character abbreviation for multiple
comparison correction method for stats::p.adjust, default is
Benjamini-Hochburg ("BH")}

\item{verbose}{(optional) TRUE or FALSE indicator for printing output to console.
default is FALSE.}

\item{save.out}{(optional) Either logical or a character string. If \code{TRUE},
it will output the result to a default file name within \code{home_dir} set in \code{initMSM()}. You can load the data with \code{x <- readRDS(file)}.
To use a non-default file name, specify a character string with the file name. It will save relative to \code{home_dir}.
There might be naming conflicts where two objects get saved to the same file. In these cases, users should specify a custom name.
default is FALSE.}

\item{x}{devMSM_histories object from \code{\link[=compareHistories]{compareHistories()}}}

\item{...}{ignored}

\item{colors}{(optional) character specifying Brewer palette or list of
colors (n(epochs)+1) for plotting (default is "Dark2" palette)}

\item{exp_lab}{(optional) character label for exposure variable in plots
(default is variable name)}

\item{out_lab}{(optional) character label for outcome variable in plots
(default is variable name)}

\item{object}{devMSM_histories object from \code{\link[=compareHistories]{compareHistories()}}}

\item{type}{Either "preds" or "comps" corresponding to the
results of \code{\link[marginaleffects:predictions]{marginaleffects::avg_predictions()}} at low and high dosages or
\code{\link[marginaleffects:comparisons]{marginaleffects::avg_comparisons()}} respectively}
}
\value{
list containing two dataframes: \code{preds} with predictions from
\code{\link[marginaleffects:predictions]{marginaleffects::avg_predictions()}} containing average expected outcome
for different exposure histories and \code{comps} with contrasts from
\code{\link[marginaleffects:comparisons]{marginaleffects::comparisons()}} comparing different exposure history
}
\description{
Takes fitted model output to created predicted values for user-specified
histories (pooling for imputed data), before conducting contrast comparisons
(pooling for imputed data), correcting for multiple comparisons, and then
plotting results.
}
\examples{
library(devMSMs)
set.seed(123)
data <- data.frame(
  ID = 1:50,
  A.1 = rnorm(n = 50),
  A.2 = rnorm(n = 50),
  A.3 = rnorm(n = 50),
  B.1 = rnorm(n = 50),
  B.2 = rnorm(n = 50),
  B.3 = rnorm(n = 50),
  C = rnorm(n = 50),
  D.3 = rnorm(n = 50)
)
obj <- initMSM(
  data,
  exposure = c("A.1", "A.2", "A.3"),
  ti_conf = c("C"),
  tv_conf = c("B.1", "B.2", "B.3", "D.3")
)
f <- createFormulas(obj, type = "short")
w <- createWeights(data = data, formulas = f)
fit <- fitModel(
  data = data, weights = w,
  outcome = "D.3", model = "m0"
)

comp <- compareHistories(
  fit = fit,
  hi_lo_cut = c(0.3, 0.6)
)
print(comp)
plot(comp)
summary(comp, "preds")
summary(comp, "comps")

comp2 <- compareHistories(
  fit = fit,
  hi_lo_cut = c(0.3, 0.6),
  reference = "l-l-l",
  comparison = c("h-h-h", "h-h-l")
)
print(comp2)
plot(comp2)
summary(comp2, "preds")
summary(comp2, "comps")

}
\seealso{
\code{\link[marginaleffects:predictions]{marginaleffects::avg_predictions()}},
\url{https://cran.r-project.org/web/packages/marginaleffects/marginaleffects.pdf};
\code{\link[marginaleffects:hypotheses]{marginaleffects::hypotheses()}},
\url{https://cran.r-project.org/web/packages/marginaleffects/marginaleffects.pdf};
\code{\link[stats:p.adjust]{stats::p.adjust()}},
\url{https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/p.adjust};
}
