% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condenseWeights.R
\name{condenseWeights}
\alias{condenseWeights}
\title{Condense weights across time points and imputed datasets}
\usage{
condenseWeights(object, weights_models)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{weights_models}{output from createWeights}
}
\value{
data_for_model_with_weights
}
\description{
Condenses weights by multiplying across time point and averaging across imputed dataset to return a single weight per exposure per person in data_for_model_with_weights
}
\examples{
condenseWeights(object, weights_models)

}
\seealso{
\code{\link[=createWeights]{createWeights()}} for more on weights_models input
}
