% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHelpers.R
\name{create_custom_comparisons}
\alias{create_custom_comparisons}
\title{Title}
\usage{
create_custom_comparisons(preds, ref_vals, comp_vals, exposure)
}
\arguments{
\item{exposure}{}
}
\description{
Title
}
