% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHelpers.R
\name{create_custom_contrasts}
\alias{create_custom_contrasts}
\title{Create custom contrasts}
\usage{
create_custom_contrasts(d, reference, comp_histories, exposure, preds)
}
\arguments{
\item{exposure}{name of exposure variable}

\item{preds}{}
}
\value{
contrasts
}
\description{
Create custom contrasts
}
