% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createForms.R
\name{createForms}
\alias{createForms}
\title{createForms}
\usage{
createForms(
  wide_long_datasets,
  covariates_to_include,
  exposures,
  outcomes,
  time_pts,
  potential_colliders = NULL
)
}
\arguments{
\item{wide_long_datasets}{from formatForWeights}

\item{covariates_to_include}{from identifyPotentialConfounds}

\item{exposures}{list of exposures of interest}

\item{outcomes}{list of outcomes of interest}

\item{time_pts}{list of time points}

\item{potential_colliders}{optional list of variables to be excluded from balancing at time point of exposure}
}
\value{
forms
}
\description{
createForms
}
