% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createForms.R
\name{createForms}
\alias{createForms}
\title{Creates formulas for generating weights}
\usage{
createForms(object, wide_long_datasets, covariates_to_include)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{covariates_to_include}{from identifyPotentialConfounds}
}
\value{
forms
}
\description{
Creates formulas for each exposure at each time point for the CBPS function that will make weights returning a list of forms
}
\examples{
createForms(object, wide_long_datasets,covariates_to_include, potential_colliders)

}
\seealso{
\code{\link[=formatForWeights]{formatForWeights()}}, \link{identifyPotentialConfounds} for more on inputs
}
