% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFormulas.R
\name{createFormulas}
\alias{createFormulas}
\alias{print.devMSM_formulas}
\title{Create balancing formulas}
\usage{
createFormulas(
  obj,
  type = c("full", "short", "update"),
  custom = NULL,
  keep_conf = NULL,
  bal_stats = NULL,
  verbose = FALSE,
  save.out = FALSE
)

\method{print}{devMSM_formulas}(x, ...)
}
\arguments{
\item{obj}{initialized MSM object from \code{\link[=initMSM]{initMSM()}}}

\item{type}{type of formula to create from 'full' (includes all lagged
time-varying confounders), 'short' (includes time-varying confounders at
t-1 lag only), or 'update' (adds to 'short' formulas any imbalanced
time-varying confounders at lags great than t-1)}

\item{custom}{(optional) custom list of formulas at each exposure time point
(default is to create automatically according to type)}

\item{keep_conf}{(optional) list of variable names reflecting confounders to
always retain in formulas (default depends on type)}

\item{bal_stats}{list of balance statistics from \code{\link[=assessBalance]{assessBalance()}}}

\item{verbose}{(optional) TRUE or FALSE indicator for printing output to console
(default is FALSE)}

\item{save.out}{(optional) TRUE or FALSE indicator to save output and
intermediary output locally (default is FALSE)}

\item{x}{devMSM_formulas object from \code{\link[=createFormulas]{createFormulas()}}}

\item{...}{ignored}
}
\value{
a list containing balancing formulas. It is the length of the number of exposure variables.
}
\description{
Creates balancing formulas relating exposure to all relevant time-varying and
time invariant confounders at each exposure time point to be used to create
IPTW weights.
}
\examples{
library(devMSMs)
data <- data.frame(
  ID = 1:50,
  A.1 = rnorm(n = 50),
  A.2 = rnorm(n = 50),
  A.3 = rnorm(n = 50),
  B.1 = rnorm(n = 50),
  B.2 = rnorm(n = 50),
  B.3 = rnorm(n = 50),
  C = rnorm(n = 50),
  D.3 = rnorm(n = 50)
)
obj <- initMSM(
  data,
  exposure = c("A.1", "A.2", "A.3"),
  ti_conf = c("C"),
  tv_conf = c("B.1", "B.2", "B.3", "D.3")
)

# Full Formulas
f <- createFormulas(obj, type = "full")
print(f)

# Short Formulas
f <- createFormulas(obj, type = "short")
print(f)

# Update Formulas
w <- createWeights(data = data, obj = obj, formulas = f)
b <- assessBalance(data = data, obj = obj, weights = w)
f <- createFormulas(obj, type = "update", bal_stats = b)
print(f)

}
