% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFormulas.R
\name{createFormulas}
\alias{createFormulas}
\title{Creates balancing formulas at each exposure time point}
\usage{
createFormulas(
  home_dir,
  exposure,
  exposure_time_pts,
  outcome,
  tv_confounders,
  ti_confounders,
  type,
  bal_stats = NULL,
  concur_conf = NULL,
  keep_conf = NULL,
  custom = NULL,
  verbose = TRUE
)
}
\arguments{
\item{home_dir}{path to home directory#' @param exposure}

\item{exposure_time_pts}{list of integers at which weights will be created/assessed that correspond to time points when exposure wass measured}

\item{outcome}{name of outcome variable with ".timepoint" suffix}

\item{tv_confounders}{list of time-varying confounders with ".timepoint" suffix}

\item{ti_confounders}{list of time invariant confounders}

\item{type}{type of formula to create from 'full', 'short', or 'update'}

\item{bal_stats}{list of balance statistics from assessBalance()}

\item{concur_conf}{list of variable names reflecting time-varying confounders to retain in formulas contemporaneously}

\item{keep_conf}{list of variable names reflecting confounders to always retain in formulas}

\item{custom}{custom list of formulas at each exposure time point}

\item{verbose}{TRUE or FALSE indicator for user output}
}
\value{
list of balancing formulas at each exposure time point
}
\description{
Creates balancing formulas at each exposure time point
}
