% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWeights.R
\name{createWeights}
\alias{createWeights}
\title{Creates IPTW balancing weights}
\usage{
createWeights(
  data,
  obj,
  formulas,
  method = c("glm", "gbm", "bart", "super", "cbps"),
  verbose = FALSE,
  save.out = FALSE,
  home_dir = NULL,
  ...
)
}
\arguments{
\item{data}{data in wide format as: a data frame, list of imputed data
frames, or mids object}

\item{obj}{initialized MSM object from \code{initMSM}}

\item{formulas}{list of balancing formulas at each time point output from
\code{createFormulas()}}

\item{method}{(optional) character string of weightitMSM() balancing method
abbreviation (default is Covariate Balancing Propensity Score "cbps")}

\item{verbose}{(optional) TRUE or FALSE indicator for user output (default is
FALSE)}

\item{save.out}{(optional) TRUE or FALSE indicator to save output and
intermediary output locally (default is FALSE)}

\item{home_dir}{path to home directory (required if 'save.out' = TRUE)}

\item{...}{pass custom arguments to \code{weightitMSM()}}
}
\value{
list of IPTW balancing weights
}
\description{
Creates IPTW balancing weights at each user-specified exposure time point
using balancing formulas that relate exposure at each time point to all
relevant confounders.
}
\examples{
library(devMSMs)
data <- data.frame(
  ID = 1:50,
  A.1 = rnorm(n = 50),
  A.2 = rnorm(n = 50),
  A.3 = rnorm(n = 50),
  B.1 = rnorm(n = 50),
  B.2 = rnorm(n = 50),
  B.3 = rnorm(n = 50),
  C = rnorm(n = 50),
  D.3 = rnorm(n = 50)
)
obj <- initMSM(
  data,
  exposure = c("A.1", "A.2", "A.3"),
  ti_conf = c("C"),
  tv_conf = c("B.1", "B.2", "B.3", "D.3")
)
f <- createFormulas(obj, type = "short")

w <- createWeights(data = data, obj = obj, formulas = f)
print(w)
plot(w)

# Methods from `WeightIt::weightitMSM`
\dontshow{if (requireNamespace("CBPS", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
w <- createWeights(data = data, obj = obj, formulas = f, method = "cbps")
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("gbm", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
w <- createWeights(data = data, obj = obj, formulas = f, method = "gbm")
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("dbarts", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
w <- createWeights(data = data, obj = obj, formulas = f, method = "bart")
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("SuperLearner", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
w <- createWeights(data = data, obj = obj, formulas = f, method = "super")
\dontshow{\}) # examplesIf}
}
\seealso{
{\code{\link[WeightIt:weightitMSM]{WeightIt::weightitMSM()}},
\url{https://ngreifer.github.io/WeightIt/reference/weightitMSM.html}}
}
