% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWeights.R
\name{createWeights}
\alias{createWeights}
\title{Creates IPTW balancing weights}
\usage{
createWeights(
  home_dir,
  data,
  exposure,
  outcome,
  formulas,
  method = "cbps",
  SL.library = NA,
  read_in_from_file = "no",
  verbose = TRUE,
  save.out = TRUE
)
}
\arguments{
\item{home_dir}{path to home directory (required if 'save.out' = TRUE)}

\item{data}{data in wide format as: a data frame, list of imputed data
frames, or mids object}

\item{exposure}{name of exposure variable}

\item{outcome}{name of outcome variable with ".timepoint" suffix}

\item{formulas}{list of balancing formulas at each time point output from
createFormulas()}

\item{method}{(optional) character string of WeightItMSM() balancing method
abbreviation (default is Covariate Balancing Propensity Score "cbps")}

\item{SL.library}{required for superLearner weighting method; see
SuperLearner::listWrappers() for options}

\item{read_in_from_file}{(optional) "yes" or "no" indicator to read in
weights that have been previously run and saved locally (default is "no")}

\item{verbose}{(optional) TRUE or FALSE indicator for user output (default is
TRUE)}

\item{save.out}{(optional) TRUE or FALSE indicator to save output and
intermediary output locally (default is TRUE)}
}
\value{
list of IPTW balancing weights
}
\description{
Creates IPTW balancing weights at each user-specified exposure time point
using balancing formulas that relate exposure at each time point to all
relevant confounders.
}
\examples{
test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),,
                    ti_confounders = "C",
                    type = "short",
                    save.out = FALSE)

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   method = "cbps",
                   save.out = FALSE)
}
\seealso{
{\code{\link[WeightIt:WeightItMSM]{WeightIt::WeightItMSM()}},
\url{https://ngreifer.github.io/WeightIt/reference/weightitMSM.html}}
}
