% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWeights.R
\name{createWeights}
\alias{createWeights}
\title{Creates balancing weights for potential confounding covariates in relation to exposure at each time point}
\usage{
createWeights(
  data,
  exposure,
  outcome,
  tv_confounders,
  formulas,
  method = "cbps",
  read_in_from_file = "no"
)
}
\arguments{
\item{read_in_from_file}{optional binary indicator of whether weights should be read in from a local file}

\item{object}{msm object that contains all relevant user inputs}

\item{wide_long_datasets}{from formatForWeights}

\item{short_forms}{from createShortForms}
}
\value{
weights_models
}
\description{
Returns a list of weights_models
}
\examples{
createWeights(object, wide_long_datasets, short_forms, read_in_from_file="no")
}
\seealso{
\code{\link[CBPS:CBPS]{CBPS::CBPS()}}, \code{\link[=formatForWeights]{formatForWeights()}}, \code{\link[=createShortForms]{createShortForms()}}
}
