% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWeights.R
\name{createWeights}
\alias{createWeights}
\alias{print.devMSM_weights}
\alias{plot.devMSM_weights}
\title{Creates IPTW balancing weights}
\usage{
createWeights(
  data,
  obj,
  formulas,
  method = c("glm", "gbm", "bart", "super", "cbps", "ipt"),
  verbose = FALSE,
  save.out = FALSE,
  ...
)

\method{print}{devMSM_weights}(x, i = 1, ...)

\method{plot}{devMSM_weights}(x, i = 1, save.out = FALSE, ...)
}
\arguments{
\item{data}{data in wide format as: a data frame, list of imputed data
frames, or \code{mids} object from the \code{mice} package}

\item{obj}{initialized MSM object from \code{\link[=initMSM]{initMSM()}}}

\item{formulas}{list of balancing formulas at each time point output from \code{\link[=createFormulas]{createFormulas()}}}

\item{method}{(optional) character string of weightitMSM() balancing method
abbreviation (default is Covariate Balancing Propensity Score "cbps")}

\item{verbose}{(optional) TRUE or FALSE indicator for printing output to console
(default is FALSE)}

\item{save.out}{(optional) Either logical or a character string. If \code{TRUE},
it will output the result to a default file name within \code{home_dir} set in \code{initMSM()}. You can load the data with \code{x <- readRDS(file)}.
To use a non-default file name, specify a character string with the file name. It will save relative to \code{home_dir}. (default is FALSE).}

\item{...}{ignored}

\item{x}{devMSM_weights object from \code{\link[=createWeights]{createWeights()}}}

\item{i}{For multiply imputed datset, \code{i} selects which imputation to print results for.
Default is \code{i = 1}. With \code{i = TRUE}, all imputed datasets will be looped over. With \code{i = NULL}, will average over all imputed datasets and summarize that.}
}
\value{
a list containing \code{\link[WeightIt:weightitMSM]{WeightIt::weightitMSM()}} output. It is the length
of the number of datasets (1 for a data.frame or the number of imputed datasets).
}
\description{
Creates IPTW balancing weights at each user-specified exposure time point
using balancing formulas that relate exposure at each time point to all
relevant confounders.
}
\examples{
library(devMSMs)
data <- data.frame(
  ID = 1:50,
  A.1 = rnorm(n = 50),
  A.2 = rnorm(n = 50),
  A.3 = rnorm(n = 50),
  B.1 = rnorm(n = 50),
  B.2 = rnorm(n = 50),
  B.3 = rnorm(n = 50),
  C = rnorm(n = 50),
  D.3 = rnorm(n = 50)
)
obj <- initMSM(
  data,
  exposure = c("A.1", "A.2", "A.3"),
  ti_conf = c("C"),
  tv_conf = c("B.1", "B.2", "B.3", "D.3")
)
f <- createFormulas(obj, type = "short")

w <- createWeights(data = data, obj = obj, formulas = f)
print(w)
plot(w)

# Methods from `WeightIt::weightitMSM`
\dontshow{if (requireNamespace("CBPS", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
w <- createWeights(data = data, obj = obj, formulas = f, method = "cbps")
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("gbm", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
w <- createWeights(data = data, obj = obj, formulas = f, method = "gbm")
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("dbarts", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
w <- createWeights(data = data, obj = obj, formulas = f, method = "bart")
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("SuperLearner", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
w <- createWeights(data = data, obj = obj, formulas = f, method = "super")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[WeightIt:weightitMSM]{WeightIt::weightitMSM()}},
\url{https://ngreifer.github.io/WeightIt/reference/weightitMSM.html};
}
