% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCovariates.R
\name{dataToImpute}
\alias{dataToImpute}
\title{Creates data structure for imputation}
\usage{
dataToImpute(
  object,
  covariates_to_include,
  exclude_covariates = NULL,
  keep_covariates = NULL
)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{covariates_to_include}{output from identifyPotentialConfounds}

\item{exclude_covariates}{list of variables to exclude based on theoretical or practical reasons}

\item{keep_covariates}{list of covariates to keep}
}
\value{
data_to_impute
}
\description{
Creates data structure for imputation with only the variables needed (i.e., exposures, outcomes, potential confounds)
}
\examples{
dataToImpute(ID, data, exposures, outcomes, covariates_to_include, exclude_covariates)

}
\seealso{
\code{\link[=formatDataStruct]{formatDataStruct()}}, \code{\link[=identifyPotenialConfounds]{identifyPotenialConfounds()}}
}
