% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCovariates.R
\name{dataToImpute}
\alias{dataToImpute}
\title{Creates data structure for imputation}
\usage{
dataToImpute(
  ID,
  data,
  exposures,
  outcomes,
  covariates_to_include,
  exclude_covariates
)
}
\arguments{
\item{ID}{person-level identifier in your dataset}

\item{data}{output from formatDataStruct}

\item{exposures}{list of variables that represent your exposures/treatments of interest}

\item{outcomes}{list of variables that represent your outcomes of interest}

\item{covariates_to_include}{output from identifyPotentialConfounds}

\item{exclude_covariates}{list of variables to exclude based on theoretical or practical reasons}
}
\value{
data_to_impute
}
\description{
Creates data structure for imputation with only the variables needed (i.e., exposures, outcomes, potential confounds)
}
\examples{
dataToImpute(ID, data, exposures, outcomes, covariates_to_include, exclude_covariates)

}
\seealso{
\code{\link[=formatDataStruct]{formatDataStruct()}}, \code{\link[=identifyPotenialConfounds]{identifyPotenialConfounds()}}
}
