% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCausalQuestion.R
\name{defineCausalQuestion}
\alias{defineCausalQuestion}
\title{Define causal questions at the heart of the subsequent marginal structural models}
\usage{
defineCausalQuestion(exposures, exposure_time_pts, outcomes, outcome_time_pts)
}
\arguments{
\item{exposures}{string of the core exposure/treatment of interest variable}

\item{exposure_time_pts}{sequence of integer time points of exposure}

\item{outcomes}{string of core outcome of interest variable}

\item{outcome_time_pts}{sequence of integer time points of outcome}
}
\description{
This code helps the individual delineate core aspects of their causal questions
}
\examples{
defineCausalQuestion(exposures, exposure_time_pts, outcomes, outcome_time_pts)

}
