% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCausalQuestion.R
\name{defineCausalQuestion}
\alias{defineCausalQuestion}
\title{Title}
\usage{
defineCausalQuestion(exposures, exposure_time_pts, outcomes, outcome_time_pts)
}
\arguments{
\item{exposure_time_pts}{sequence of integer time points of exposure}

\item{outcome_time_pts}{sequence of integer time points of outcome}

\item{exposure}{string of the core exposure/treatment of interest variable}

\item{outcome}{string of core outcome of interest variable}
}
\description{
Title
}
