% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devMSMs-package.R
\docType{package}
\name{devMSMs-package}
\alias{devMSMs}
\alias{devMSMs-package}
\title{devMSMs: Tools for Conducting Marginal Structural Models with Developmental Data}
\description{
Functions for preparing, assessing, and implementing MSMS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://istallworthy.github.io/devMSMs/}
}

}
\author{
\strong{Maintainer}: Isabella Stallworthy \email{istall@seas.upenn.edu}

Authors:
\itemize{
  \item Meriah DeJoseph \email{meriahd@stanford.edu}
  \item Emily Padrutt \email{padru004@umn.edu}
  \item Noah Greifer \email{ngreifer@iq.harvard.edu}
  \item Daniel Berry \email{dberry@umn.edu}
}

}
\keyword{internal}
