% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.avg_imp_bal_stats}
\alias{.avg_imp_bal_stats}
\title{Averages balance stats across imputed datasets: df of all times}
\usage{
.avg_imp_bal_stats(bal_stats)
}
\description{
Averages balance stats across imputed datasets: df of all times
}
\keyword{internal}
