% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_hist.R
\name{eval_hist}
\alias{eval_hist}
\title{Visualize distribution of sample across exposure histories}
\usage{
eval_hist(
  data,
  exposure,
  tv_confounders,
  epochs,
  time_pts,
  hi_lo_cut,
  ref,
  comps
)
}
\arguments{
\item{exposure}{name of exposure variable}

\item{tv_confounders}{list of time-varying confounders with ".timepoint" suffix}

\item{hi_lo_cut}{list of two numbers indicating quantile values that reflect high and low values, respectively, for continuous exposure}

\item{comps}{}
}
\value{
none
}
\description{
Visualize distribution of sample across exposure histories
}
