% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fitModel}
\alias{fitModel}
\title{Fit weighted model
This code fits a weighted marginal structural model to examine the effects of different exposure histories on outcomes}
\usage{
fitModel(
  home_dir,
  ID,
  exposures,
  exposure_epochs,
  outcomes,
  outcome_time_pts,
  data_for_model_with_weights_cutoff,
  unbalanced_covariates_for_models
)
}
\arguments{
\item{home_dir}{path to home directory for the project}

\item{ID}{person-level identifier in your dataset}

\item{exposures}{list of variables that represent your exposures/treatments of interest}

\item{exposure_epochs}{data frame containing epochs and correponding time points for exposure histories}

\item{outcomes}{list of variables that represent your outcomes of interest}

\item{outcome_time_pts}{list of time points for outcomes}

\item{data_for_model_with_weights_cutoff}{dataset with truncated weights see truncateWeights}

\item{unbalanced_covariates_for_models}{unbalanced covariates see assessBalance}
}
\value{
marginal structural models
}
\description{
Fit weighted model
This code fits a weighted marginal structural model to examine the effects of different exposure histories on outcomes
}
\examples{
fitModel(ID, exposures, exposure_epochs, outcomes, outcome_time_pts, data_for_model_with_weights_cutoff, unbalanced_covariates_for_models, hi_cutoff=75, lo_cutoff=25)
}
\seealso{
\code{\link[=truncateWeights]{truncateWeights()}}, \code{\link[=asesssBalance]{asesssBalance()}}
}
