% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fitModel}
\alias{fitModel}
\title{Fit weighted model
This code fits a weighted marginal structural model to examine the effects of different exposure histories on outcome}
\usage{
fitModel(
  object,
  data_for_model_with_weights_cutoff,
  unbalanced_covariates_for_models = NULL,
  model = "m4"
)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{data_for_model_with_weights_cutoff}{dataset with truncated weights see truncateWeights}

\item{unbalanced_covariates_for_models}{unbalanced covariates see assessBalance}
}
\value{
marginal structural models
}
\description{
Fit weighted model
This code fits a weighted marginal structural model to examine the effects of different exposure histories on outcome
}
\examples{
fitModel(object, data_for_model_with_weights_cutoff, unbalanced_covariates_for_models)
}
\seealso{
\code{\link[=truncateWeights]{truncateWeights()}}, \code{\link[=asesssBalance]{asesssBalance()}}
}
