% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fitModel}
\alias{fitModel}
\title{Fit outcome model}
\usage{
fitModel(
  home_dir,
  data,
  weights,
  exposure,
  exposure_time_pts,
  outcome,
  tv_confounders,
  model,
  family = gaussian,
  link = "identity",
  int_order = NA,
  covariates = NULL,
  epochs = NULL,
  verbose = TRUE,
  save.out = TRUE
)
}
\arguments{
\item{home_dir}{path to home directory}

\item{data}{data in wide format as: a data frame, list of imputed
data frames, or mids object}

\item{weights}{list of IPTW weights output from createWeights()}

\item{exposure}{name of exposure variable}

\item{exposure_time_pts}{list of integers at which weights will be
created/assessed that correspond to time points when exposure was measured}

\item{outcome}{name of outcome variable with ".timepoint" suffix}

\item{tv_confounders}{list of time-varying confounders with ".timepoint"
suffix}

\item{model}{character indicating one of the following outcome models:
\itemize{
\item "m0" (exposure main effects)
\item "m1" (exposure main effects & covariates)
\item "m2" (exposure main effects & their interactions)
\item "m3" (exposure main effects, their interactions, & covariates)
}}

\item{family}{(optional) family function specification for svyglm model}

\item{link}{(optional) character link function specification for svyglm model}

\item{int_order}{integer specification of highest order exposure main effects
interaction, required for interaction models ("m2", "m3")}

\item{covariates}{list of characters reflecting variable names of covariates,
required for covariate models ("m1", "m3")}

\item{epochs}{(optional) data frame of exposure epoch labels and values}

\item{verbose}{(optional) TRUE or FALSE indicator for user output (default is
TRUE)}

\item{save.out}{(optional) TRUE or FALSE indicator to save output and intermediary output locally (default is TRUE)}
}
\value{
list of svyglm model output
}
\description{
Fits weighted marginal outcome model as a generalized linear model of the
user's choosing, relating exposure main effects to outcome using IPTW
weights.
}
\seealso{
{\code{\link[survey:svyglm]{survey::svyglm()}} for more on family/link specifications, <url1>}

{\code{\link[=createWeights]{createWeights()}}, <url1>}
}
