% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fitModel}
\alias{fitModel}
\title{Fit weighted model
This code fits a weighted marginal structural model to examine the effects of different exposure histories on outcome}
\usage{
fitModel(
  home_dir,
  data,
  weights,
  exposure,
  exposure_time_pts,
  outcome,
  tv_confounders,
  model,
  family = gaussian,
  link = "identity",
  int_order = NA,
  covariates = NULL,
  epochs = NULL,
  user.o = TRUE
)
}
\arguments{
\item{model}{user-specified model}
}
\value{
fits

fits
}
\description{
Fit weighted model
This code fits a weighted marginal structural model to examine the effects of different exposure histories on outcome
}
