% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessDataStruct.R
\name{formatDataStruct}
\alias{formatDataStruct}
\title{assessDataStruct}
\usage{
formatDataStruct(data_path, home_dir, missing, factor_covariates)
}
\arguments{
\item{data_path}{path to cleaned dataset}

\item{home_dir}{path to home directory for the project}

\item{factor_covariates}{list of covariates that are factors}
}
\value{
data formatted dataset
}
\description{
assessDataStruct
}
