% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatForWeights.R
\name{formatForWeights}
\alias{formatForWeights}
\title{Format data for creating balancing weights}
\usage{
formatForWeights(object, data, imputed_datasets)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{imputed_datasets}{output from imputeData}
}
\value{
wide_long_datasets
}
\description{
This code formats the imputed datasets for calculating balancing weights
}
\examples{
formatForWeights(object, data, imputed_datasets)

}
\seealso{
\code{\link[=formatDataStruct]{formatDataStruct()}}, \code{\link[=imputeData]{imputeData()}}
}
