% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatForWeights.R
\name{formatForWeights}
\alias{formatForWeights}
\title{Format data for creating balancing weights}
\usage{
formatForWeights(object, data, imputed_datasets = list(), just_imputed = "yes")
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{imputed_datasets}{output from imputeData}

\item{just_imputed}{"yes"= you have imputed datasets in global environment or "no" but they are saved locally from previous run}
}
\value{
wide_long_datasets
}
\description{
This code formats the imputed datasets for calculating balancing weights using CBPS and reutrns a list of wide/long datasets and creates dataset for future modeling
}
\examples{
formatForWeights(object, data, imputed_datasets=list(), just_imputed="yes")

}
\seealso{
\code{\link[=formatDataStruct]{formatDataStruct()}}, \code{\link[=imputeData]{imputeData()}}
}
