% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHelpers.R
\name{get_comparison_values}
\alias{get_comparison_values}
\title{Finds custom comparison values}
\usage{
get_comparison_values(d, comp_histories)
}
\arguments{
\item{d}{data frame of high and low values per exposure main effect}

\item{comp_histories}{comparison sequence(s) of "h" and/or "l" (e.g., "h-h-h")}
}
\value{
comparison values
}
\description{
Finds custom comparison values
}
\examples{
d <- data.frame(e = c("A.1", "A.2", "A.3"),
               l = c(0, 0, 0),
               h = c(1, 1, 1))
r <- get_comparison_values(d = d,
                          comp_histories = "l-l-l" )
r <- get_comparison_values(d = d,
                          comp_histories = "h-h-h" )
r <- get_comparison_values(d = d,
                         comp_histories = c("h-h-h", "h-h-l"))
}
