% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHelpers.R
\name{get_reference_values}
\alias{get_reference_values}
\title{Finds custom reference values}
\usage{
get_reference_values(d, reference)
}
\arguments{
\item{reference}{}
}
\value{
reference values
}
\description{
Finds custom reference values
}
