% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCovariates.R
\name{identifyCovariates}
\alias{identifyCovariates}
\title{Identifies the available covariates}
\usage{
identifyCovariates(data, ID, exposures, outcomes, exclude_covariates = NULL)
}
\arguments{
\item{data}{output from formatDataStruct}

\item{ID}{person-level identifier in your dataset}

\item{exposures}{list of variables that represent your exposures/treatments of interest}

\item{outcomes}{list of variables that represent your outcomes of interest}

\item{exclude_covariates}{list of variables to exclude based on theoretical or practical reasons}
}
\value{
potential_covariates
}
\description{
Identifies the covariates in in the dataset that will be candidates for confounding
}
\details{
identifyCovariates
}
\examples{
identifyCovariates(data, ID, exposures, outcomes, exclude_covariates)

}
\seealso{
\code{\link[=formatDataStruct]{formatDataStruct()}}
}
