% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCovariates.R
\name{identifyCovariates}
\alias{identifyCovariates}
\title{Identifies the available covariates}
\usage{
identifyCovariates(object, data)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{data}{data from formatDataStruct}
}
\value{
potential_covariates
}
\description{
Identifies the covariates in in the dataset that will be candidates for confounding
}
\details{
identifyCovariates
}
\examples{
identifyCovariates(object,data, exclude_covariates)

}
\seealso{
\code{\link[=formatDataStruct]{formatDataStruct()}}
}
