% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCovariates.R
\name{identifyPotentialConfounds}
\alias{identifyPotentialConfounds}
\title{Identify potential confounds}
\usage{
identifyPotentialConfounds(
  home_dir,
  data,
  time_pt_datasets,
  time_pts,
  exclude_covariates = NULL
)
}
\arguments{
\item{home_dir}{path to home directory for the project}

\item{data}{output from formatDataStruct}

\item{time_pt_datasets}{output from makeTimePtDatasets}

\item{time_pts}{list of time points along your developmental path of interest for which you have at least one measurement}

\item{exclude_covariates}{list of variables to exclude based on theoretical or practical reasons}
}
\value{
covariates_to_include
}
\description{
Identifies potential confounds from the covariates based on those that are correlated with either exposures or outcomes above 0.1
}
\examples{
identifyPotentialConfounds(home_dir, data, time_pt_datasets, time_pts, exclude_covariates=NULL)

}
\seealso{
\code{\link[=formatDataStruct]{formatDataStruct()}}, \code{\link[=makeTimePtDatasets]{makeTimePtDatasets()}}
}
