% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeData.R
\name{imputeData}
\alias{imputeData}
\title{Imputes dataset so there is no missing at each time point using parallel processing to speed up}
\usage{
imputeData(
  data,
  m = 5,
  method = "rf",
  home_dir,
  exposure,
  outcome,
  tv_confounders,
  ti_confounders,
  read_imps_from_file = "no"
)
}
\arguments{
\item{data}{data in wide format}

\item{m}{integer number of imputed datasets}

\item{method}{character string of imputation method from mice()}

\item{home_dir}{path to home directory}

\item{exposure}{name of exposure variable}

\item{outcome}{name of outcome variable with ".timepoint" suffix}

\item{tv_confounders}{list of time-varying confounders with ".timepoint" suffix}

\item{ti_confounders}{list of time invariant confounders}

\item{read_in_from_file}{"yes" or "no" indicatorto read in weights that have been previously run and saved locally}
}
\value{
mice object of m imputed datasets
}
\description{
Imputes dataset so there is no missing at each time point using parallel processing to speed up
}
