% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeData.R
\name{imputeData}
\alias{imputeData}
\title{Imputes dataset so there is no missing at each time point}
\usage{
imputeData(
  object,
  data_to_impute,
  max.resample = 100,
  cs = NULL,
  priors = NULL,
  lags = NULL,
  intercs = FALSE,
  leads = NULL,
  splinetime = NULL,
  logs = NULL,
  sqrts = NULL,
  lgstc = NULL,
  noms = NULL,
  bounds = NULL
)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{data_to_impute}{output from dataToImpute}

\item{max.resample}{Amelia input for how many times amelia should redraw imputed values when trying to meet logical constraints of obounds}

\item{cs}{Amelia input for col num or var indicating cross section variable}

\item{priors}{Amelia input col matrix indicateing priors for indiv missing obs or variable-wide missing}

\item{lags}{Amelia input for cols that should have lags included in imputation model}

\item{intercs}{Amelia input logical var indicating if time effects of polytime should vary across cross-section}

\item{leads}{Amelia input for cols that should have leads included in imputation model}

\item{splinetime}{Amelia input integer of 0+ to control cubic smoothing splines of time}

\item{logs}{Amelia input for cols that require log-linear transformation}

\item{sqrts}{Amelia input for cols that require sqrt transformation}

\item{lgstc}{Amelia input for cols that require lgstc transformation}

\item{noms}{Amelia input for cols that are nominal}

\item{bounds}{Amelia input for 3-col matrix for logical bounds on imputations}
}
\value{
imputed_datasets imputation results
}
\description{
Creates m imputed datasets from original dataset using the Aeelia package. See Amelia documentation for more detail: https://cran.r-project.org/web/packages/Amelia/Amelia.pdf
}
\examples{
imputeData(object, data_to_impute, max.resample = 100, cs=NULL, priors=NULL, lags=NULL, intercs=FALSE, leads=NULL, splinetime=NULL, logs=NULL, sqrts=NULL, lgstc=NULL, noms=NULL, bounds=NULL)

}
\seealso{
\code{\link[=dataToImput]{dataToImput()}}, \code{\link[Amelia:amelia]{Amelia::amelia()}}
}
