% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfimputeData.R, R/imputeData.R
\name{imputeData}
\alias{imputeData}
\title{Imputes dataset so there is no missing at each time point}
\usage{
imputeData(object, data_to_impute, read_imps_from_file = "no")

imputeData(object, data_to_impute, read_imps_from_file = "no")
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{data_to_impute}{output from dataToImpute}
}
\value{
imputed_datasets imputation results

imputed_datasets imputation results
}
\description{
Creates m imputed datasets from original dataset using the Aeelia package. See Amelia documentation for more detail: https://cran.r-project.org/web/packages/Amelia/Amelia.pdf

Creates m imputed datasets from original dataset using the Aeelia package. See Amelia documentation for more detail: https://cran.r-project.org/web/packages/Amelia/Amelia.pdf
}
\examples{
imputeData(object, data_to_impute, read_imps_from_file="no", max.resample = 100, cs=NULL, priors=NULL, lags=NULL, intercs=FALSE, leads=NULL, splinetime=NULL, logs=NULL, sqrts=NULL, lgstc=NULL, noms=NULL, bounds=NULL)

imputeData(object, data_to_impute, read_imps_from_file="no", max.resample = 100, cs=NULL, priors=NULL, lags=NULL, intercs=FALSE, leads=NULL, splinetime=NULL, logs=NULL, sqrts=NULL, lgstc=NULL, noms=NULL, bounds=NULL)

}
\seealso{
\code{\link[=dataToImpute]{dataToImpute()}}, \code{\link[mice:mice]{mice::mice()}}

\code{\link[=dataToImpute]{dataToImpute()}}, \code{\link[mice:mice]{mice::mice()}}
}
