% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeData.R
\name{imputeData}
\alias{imputeData}
\title{Imputes dataset so there is no missing at each time point using parallel
processing to speed up}
\usage{
imputeData(
  data,
  m = 5,
  method = "rf",
  home_dir,
  exposure,
  outcome,
  para_proc = TRUE,
  read_imps_from_file = "no"
)
}
\arguments{
\item{data}{data in wide format}

\item{m}{(optional) integer number of imputed datasets (default is 5)}

\item{method}{(optional) character string of imputation method from mice()
(default is random forest "rf")}

\item{home_dir}{path to home directory}

\item{exposure}{name of exposure variable}

\item{outcome}{name of outcome variable with ".timepoint" suffix}

\item{para_proc}{(optional) TRUE/FALSE whether to do parallel processing
using multiple cores to speed up process (default = TRUE)}

\item{read_in_from_file}{(optional) "yes" or "no" indicatorto read in weights
that have been previously run and saved locally (default is "no")}
}
\value{
mice object of m imputed datasets
}
\description{
Imputes dataset so there is no missing at each time point using parallel
processing to speed up
}
\seealso{
{\code{\link[mice:mice]{mice::mice()}},
\url{https://cran.r-project.org/web/packages/mice/index.html}}
}
