% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMSM.R
\name{initMSM}
\alias{initMSM}
\title{Initial step in \code{devMSMs} workflow}
\usage{
initMSM(data, exposure, tv_conf, ti_conf)
}
\arguments{
\item{data}{data in wide format as: a data frame, list of imputed data
frames, or mids object}

\item{exposure}{names of exposure variables with ".timepoint" suffix}

\item{tv_conf}{list of time-varying confounders with ".timepoint"
suffix, should include exposure and outcome variables (at least
time-varying exposure variables required here)}

\item{ti_conf}{list of time invariant confounders (at least one required)}
}
\value{
object of class \code{devMSM} that contains the initialized information.
}
\description{
Initial step in \code{devMSMs} workflow
}
\details{
By \code{.timepoint} suffix, we mean that time-varying and exposure variable
names must end with either \code{.#} or a \verb{_#}. This allows us to extract the
time-period when the variable is measured and will allow us to properly
create the formulae (omitting future mediators)
}
