% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_love_plot.R
\name{make_love_plot}
\alias{make_love_plot}
\title{Create love plots showing balancing statistics}
\usage{
make_love_plot(
  balance_stats,
  exposure_type = c("continuous", "binary"),
  k = 0,
  weight_method = NULL
)
}
\arguments{
\item{balance_stats}{data frame of balance statistics}

\item{exposure_type}{String either "continuous" or "binary"}

\item{k}{imputation number}

\item{weight_method}{name of weighting method (NULL for unweighted)}
}
\value{
ggplot2 object
}
\description{
Create love plots showing balancing statistics
}
\examples{
print("tbd")

}
