% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMplusInputs.R
\name{makeMplusInputs}
\alias{makeMplusInputs}
\title{Create MPlus inputs for running weighted growth curve models
Code for assessing effects of exposure histories with repeated outcomes}
\usage{
makeMplusInputs(object, mplusObject)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{covariates}{any covariates you wish to include in the model}
}
\value{
none
}
\description{
Create MPlus inputs for running weighted growth curve models
Code for assessing effects of exposure histories with repeated outcomes
}
\examples{
makeMplusInputs(object, mplusObject)
}
