% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessDataStruct.R
\name{makeTimePtDatasets}
\alias{makeTimePtDatasets}
\title{Makes timepoint datasets}
\usage{
makeTimePtDatasets(object, data)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{data}{output from formatDataStruct}
}
\description{
Creates datasets unique to each time point for future use and returns a list of time point datasets
}
\examples{
makeTimePtDatasets(object, data)
}
\seealso{
\code{\link[=fomatDataStruct]{fomatDataStruct()}}
}
