% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcCorrection.R
\name{mcCorrection}
\alias{mcCorrection}
\title{Multiple comparison correction
Adjusts p-values for multiple comparisons by exposure-outcome pairing}
\usage{
mcCorrection(object, history_comparisons)
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{history_comparisons}{output from compareHistories}
}
\value{
significant_comparisons
}
\description{
Multiple comparison correction
Adjusts p-values for multiple comparisons by exposure-outcome pairing
}
\examples{
mcCorrection(object, history_comparisons)
}
\seealso{
\code{\link[stats:p.adjust]{stats::p.adjust()}} \code{\link[=comparisonHistories]{comparisonHistories()}}
}
