% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplusObject.R
\name{mplusObject}
\alias{mplusObject}
\title{Creates mplus object
Gather information for creating Mplus input template files}
\usage{
mplusObject(data_file, covariates, reference)
}
\arguments{
\item{data_file}{location of wide/long .csv file with outcomes, exposures, covariates, and weights for modeling}

\item{covariates}{data frame specifying the names of outputs as well as any technical covariates for modeling}

\item{reference}{reference history for comparison to all other histories}
}
\value{
object
}
\description{
Creates mplus object
Gather information for creating Mplus input template files
}
\examples{
mplusObject(data_file, covariates, reference, hi_cutoff=0.75, lo_cutoff=0.25)
}
