% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmObject.R
\name{msmObject}
\alias{msmObject}
\title{Create msm object
Gathers all information needed for the functions and checks user inputs}
\usage{
msmObject(
  data_file,
  home_dir,
  ID,
  time_pts,
  time_var,
  missing,
  time_varying_variables = NULL,
  continuous_variables = NULL,
  factor_covariates = NULL,
  m = 5,
  imp_method = "cart",
  exposure,
  exposure_time_pts,
  short_form_lag = 1,
  weights_method = "cbps",
  balance_thresh = 0.1,
  weights_percentile_cutoff = 0.95,
  exposure_epochs,
  reference = NA,
  comparisons = NA,
  hi_cutoff,
  lo_cutoff,
  mc_method = "BH",
  outcome,
  outcome_time_pt,
  keep_concurrent_tv_vars = NULL,
  time_var_exclude = NULL,
  exposure_labels = NULL,
  outcome_labels = NULL,
  dose_level = "h",
  colors = "Dark2"
)
}
\arguments{
\item{home_dir}{path to home directory for the project}

\item{ID}{person-level identifier in your dataset}

\item{time_pts}{list of time points along your developmental path of interest for which you have at least one measurement}

\item{time_var}{variable in your long dataset that designates developmental time}

\item{missing}{missing data marker in your dataset}

\item{time_varying_variables}{all variables in dataset that are time-varying}

\item{continuous_variables}{list of continuous variables in your dataset}

\item{factor_covariates}{all variables that are factors}

\item{m}{number of imputations}

\item{exposure_time_pts}{times when exposures occurred}

\item{balance_thresh}{correlation value above which covariates are not considered balanced with respect to exposure}

\item{weights_percentile_cutoff}{percentile value for cutting off and replacing heavy tails of weights}

\item{exposure_epochs}{data frame containing epochs and correponding time points for exposure histories}

\item{outcome}{list of variables that represent your outcome of interest}

\item{time_var_exclude}{time-varying variables that should not be present because of planned missingness}

\item{colors}{colors for plotting dose}

\item{data_path}{path to cleaned dataset}

\item{exposures}{list of variables that represent your exposures/treatments of interest}

\item{outcome_time_pts}{list of time points for outcome}
}
\description{
Create msm object
Gathers all information needed for the functions and checks user inputs
}
\examples{
msmObject(data_file, home_dir, ID, time_pts, time_var, missing, time_varying_variables=NULL, continuous_variables=NULL, factor_covariates=NULL, m=5, imp_method="cart", exposure, exposure_time_pts, short_form_lag=1, weights_method="cbps", balance_thresh=0.1, weights_percentile_cutoff=0.95, exposure_epochs, reference=NA, comparisons=NA, hi_cutoff=.75,lo_cutoff=.25, mc_method="BH", outcome, outcome_time_pt, keep_concurrent_tv_vars=NULL, time_var_exclude=NULL,  exposure_labels=NULL, outcome_labels=NULL, dose_level="h", colors="Dark2")
}
