% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmObject.R
\name{msmObject}
\alias{msmObject}
\title{Create msm object
Gathers all information needed for the functions}
\usage{
msmObject(
  data_path,
  home_dir,
  ID,
  time_pts,
  time_var,
  missing,
  time_varying_variables = NULL,
  continuous_variables = NULL,
  factor_covariates = NULL,
  m = 5,
  exposures,
  exposure_time_pts,
  balance_thresh = 0.12,
  weights_percentile_cutoff = 0.9,
  exposure_epochs,
  reference = "",
  comparisons = "",
  hi_cutoff = 0.75,
  lo_cutoff = 0.25,
  mc_method = "BH",
  outcomes,
  outcome_time_pt,
  mandatory_keep_covariates = NULL,
  exclude_covariates = NULL,
  time_var_exclude = NULL,
  potential_colliders = NULL,
  exposure_labels = NULL,
  outcome_labels = NULL,
  colors = "Dark2"
)
}
\arguments{
\item{data_path}{path to cleaned dataset}

\item{home_dir}{path to home directory for the project}

\item{ID}{person-level identifier in your dataset}

\item{time_pts}{list of time points along your developmental path of interest for which you have at least one measurement}

\item{time_var}{variable in your long dataset that designates developmental time}

\item{missing}{missing data marker in your dataset}

\item{time_varying_variables}{all variables in dataset that are time-varying}

\item{continuous_variables}{list of continuous variables in your dataset}

\item{factor_covariates}{all variables that are factors}

\item{m}{number of imputations}

\item{exposures}{list of variables that represent your exposures/treatments of interest}

\item{exposure_time_pts}{times when exposures occurred}

\item{balance_thresh}{correlation value above which covariates are not considered balanced with respect to exposure}

\item{weights_percentile_cutoff}{percentile value for cutting off and replacing heavy tails of weights}

\item{exposure_epochs}{data frame containing epochs and correponding time points for exposure histories}

\item{outcomes}{list of variables that represent your outcomes of interest}

\item{time_var_exclude}{time-varying variables that should not be present because of planned missingness}

\item{colors}{colors for plotting dose}

\item{outcome_time_pts}{list of time points for outcomes}
}
\description{
Create msm object
Gathers all information needed for the functions
}
\examples{
msmObject(data_path, home_dir, ID, time_pts, time_var, missing, m, exposures, exposure_time_pts, exposure_epochs, outcomes, outcome_time_pt, continuous_variables, time_var_exclude)
}
