% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareHelpers.R
\name{perform_multiple_comparison_correction}
\alias{perform_multiple_comparison_correction}
\title{Conduct multiple comparison correction}
\usage{
perform_multiple_comparison_correction(
  comps,
  reference,
  comp_histories,
  method
)
}
\arguments{
\item{method}{character abbreviation for multiple comparison correction method}
}
\value{
comparison table with corrected p-values
}
\description{
Conduct multiple comparison correction
}
