% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotResults}
\alias{plotResults}
\title{Plot results from history comparisons
Code to plot predicted values of the different exposure histories by history and colored by dosage of exposure}
\usage{
plotResults(object, exposure_labels = "", outcome_labels = "")
}
\arguments{
\item{object}{msm object that contains all relevant user inputs}

\item{exposure_labels}{optional list of labels for exposure variables for plotting purposes}

\item{outcome_labels}{optional list of labels for outcome variables for plotting purposes}

\item{best_models}{outcome from assessModels}
}
\value{
plots
}
\description{
Plot results from history comparisons
Code to plot predicted values of the different exposure histories by history and colored by dosage of exposure
}
\examples{
plotResults(object, best_models, exposure_labels="", outcome_labels="")
}
\seealso{
\code{\link[=assessModels]{assessModels()}}
}
