% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotResults}
\alias{plotResults}
\title{Plot results from history comparisons
Code to plot predicted values of the different exposure histories by history and colored by dosage of exposure}
\usage{
plotResults(msm_object, history_estimates)
}
\arguments{
\item{msm_object}{msm object that contains all relevant user inputs}

\item{history_estimates}{pooled predicted values for the marginal model}
}
\value{
plots
}
\description{
Plot results from history comparisons
Code to plot predicted values of the different exposure histories by history and colored by dosage of exposure
}
\examples{
plotResults(msm_object, history_estimates)
}
