% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateWeights.R
\name{truncateWeights}
\alias{truncateWeights}
\title{truncateWeights}
\usage{
truncateWeights(
  data_for_model_with_weights,
  home_dir,
  exposures,
  percentile_cutoff = 0.9
)
}
\arguments{
\item{data_for_model_with_weights}{output from condenseWeights}

\item{home_dir}{path to home directory for the project}

\item{exposures}{list of variables that represent your exposures/treatments of interest}

\item{percentile_cutoff}{percentile value for cutting off and replacing heavy tails of weights}
}
\value{
data_for_model_with_weights_cutoff
}
\description{
truncateWeights
}
