% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateWeights.R
\name{truncateWeights}
\alias{truncateWeights}
\title{Truncates weights}
\usage{
truncateWeights(
  data_for_model_with_weights,
  home_dir,
  exposures,
  outcomes,
  percentile_cutoff = 0.9
)
}
\arguments{
\item{data_for_model_with_weights}{output from condenseWeights}

\item{home_dir}{path to home directory for the project}

\item{exposures}{list of variables that represent your exposures/treatments of interest}

\item{percentile_cutoff}{percentile value for cutting off and replacing heavy tails of weights}
}
\value{
data_for_model_with_weights_cutoff
}
\description{
Code to 'cut off' weights at 90th percentile and populate all of those above at 90th percentile value to avoid heavy tails that can bias results
}
\examples{
truncateWeights(data_for_model_with_weights, home_dir, exposures, percentile_cutoff=0.90)

}
\seealso{
\code{\link[=condenseWeights]{condenseWeights()}}
}
