% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateWeights.R
\name{truncateWeights}
\alias{truncateWeights}
\title{Truncates weights}
\usage{
truncateWeights(data_for_model_with_weights, msm_object)
}
\arguments{
\item{data_for_model_with_weights}{output from condenseWeights}

\item{msm_object}{msm object that contains all relevant user inputs}
}
\value{
dat_w_t list of truncated weights for each cutoff value
}
\description{
Code to 'cut off' weights at the 90th percentile and populate all of those above the 90th percentile value to avoid heavy tails that can bias results.
}
\examples{
truncateWeights(object, data_for_model_with_weights)

}
